% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_in_silico_root.R
\name{add_in_silico_root}
\alias{add_in_silico_root}
\title{build an in-silico root cell and clone profile}
\usage{
add_in_silico_root(cnr, base.ploidy = 2L, cell.name = "diploid", female = TRUE)
}
\arguments{
\item{cnr}{a cnr}

\item{base.ploidy}{integer, base ploidy (2, 4, 6, 8), default 2, for diploid cell}

\item{cell.name}{name of root cell. default "diploid"}

\item{female}{logical, weather to build a female or male
 cell and clone, default is female. If female is NULL, no sex chromosomes
correction is performed, all bins will have copy number equal base.ploidy}
}
\value{
Append an in-silico diplod (or polyploid) cell to the cnr.  By default a female
human genome is created.  If FALSE, a male human genome is constructed.  NULL will
not peform a gender correction and the entire genome will be set to the base ploidy.

In tetraploid males copy number is half of the base.ploidy e.g. for 4n : X = 2, Y = 2.
}
\description{
By default cell is a female diploid cell
}
\examples{

data(cnr)

cnr <- add_in_silico_root(cnr)

head(cnr$Y)

}
