% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addInfo.R
\name{addBinInfo}
\alias{addBinInfo}
\title{Adds new columns to the chromInfo}
\usage{
addBinInfo(cnr, df)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{df}{a data.frame with the data to incorporate.  Particularly
useful for adding p-values, genetic effects, etc to the bins}
}
\value{
Returns a CNR object with added columns to the chromInfo. e.g. p-values for genome wide scans
}
\description{
The function adds information at the bin level.
}
\examples{

data(cnr)

fakePval <- data.frame(pval = runif(5000))

cnr <- addBinInfo(cnr, df = fakePval)

head(cnr$chromInfo)

}
