% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addInfo.R
\name{addGeneInfo}
\alias{addGeneInfo}
\title{Adds Gene Information to the gene.index}
\usage{
addGeneInfo(
  cnr,
  df,
  gene.ids = "hgnc.symbol",
  sort = FALSE,
  all.index = TRUE,
  ...
)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{df}{a data.frame with the data to incorporate.  Particularly
useful for gene annotations, filters e.g. OncoKB, or results from
analyses e.g. p-values, genetic effects, etc to the genes}

\item{gene.ids}{column for gene.ids, must be included in both}

\item{sort}{wether to sort the ouput object, default is FALSE}

\item{all.index}{logical, use all of gene.index. When FALSE, return only
the intersect of gene.idex and df.  default TRUE}

\item{...}{additional parameters passed to \code{\link[base]{merge}}}
}
\value{
Returns a CNR object with added columns from df to the gene.index.
   e.g. p-values for genome wide scans.

By default merge occurs using the intersect of column names of
the gene.index and df.  To change this, use by.x and by.y.  See
 \code{\link[base]{merge}}.  Rownames are added after the merge
 from the column specified as gene.ids
}
\description{
The function adds information to the gene.index.#'
}
\examples{
data(cnr)

fakeGenePval <- data.frame(hgnc.symbol = cnr$gene.index$hgnc.symbol,
                       pval = runif(nrow(cnr$gene.index)))

cnr <- addGeneInfo(cnr, df = fakeGenePval)

head(cnr$gene.index)

}
