% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addInfo.R
\name{addInfo}
\alias{addInfo}
\title{Adds new columns to the chromInfo and gene.index}
\usage{
addInfo(cnr, df, gdf = NULL, ...)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{df}{a data.frame with the data to incorporate into chromInfo.  Particularly
useful for adding p-values, genetic effects, etc to the bins}

\item{gdf}{a data.frame with the data to incorporate into gene.index.  Particularly
useful for adding p-values, genetic effects, etc to the genes}

\item{...}{other parameters passed to \code{\link{addGeneInfo}}, subsequently to
\code{\link[base]{merge}}}
}
\value{
Returns a CNR object with added columns to the chromInfo and to the gene.index.
e.g. p-values for genome wide scans,
}
\description{
The function adds information at the bin and gene level by running \code{\link{addBinInfo}},
and \code{\link{addGeneInfo}}
}
\examples{

data(cnr)

fakeBinPval <- data.frame(pval = runif(5000))

fakeGenePval <- data.frame(hgnc.symbol = cnr$gene.index$hgnc.symbol,
pval = runif(nrow(cnr$gene.index)))
                       
cnr <- addInfo(cnr, df = fakeBinPval, gdf = fakeGenePval)

head(cnr$chromInfo)
head(cnr$gene.index)

## add info only to bins

fakeBinPval2 <- data.frame(pval = runif(5000))
cnr <- addInfo(cnr, df = fakeBinPval2)

head(cnr$chromInfo)
head(cnr$gene.index)

}
