% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPheno.R
\name{addPheno}
\alias{addPheno}
\title{Adds additional phenotypes for a new trait}
\usage{
addPheno(cnr, df, sort = FALSE, ...)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{df}{a data frame with new Y traits}

\item{sort}{logical, sort data.frame after merge \code{\link[base]{merge}}}

\item{...}{additional arguments by merge}
}
\value{
Returns a CNR object with added columns to Y, e.g. additional cell
 phenotypes, histologies, etc.
}
\description{
add phenotype columns to the data
}
\examples{

data(cnr)

rand3 <- data.frame(cellID = cnr$Y$cellID,
                    rand3 = rnorm(nrow(cnr$Y), mean = 2, sd = 1))
                    
cnr <- addPheno(cnr, df = rand3, by = "cellID")

}
