% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addQC.R
\name{addQC}
\alias{addQC}
\title{addQC}
\usage{
addQC(cnr, df, ...)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{df}{data containg the additional metadata to be incorporated into the qc}

\item{...}{additional parameters for sort}
}
\value{
returns a CNR bundle with new QC columns
}
\description{
Add QC columns to the qc data frame
}
\examples{

data(cnr)

mapd <- data.frame(t(apply(cnr$X, 2, mapd)))
mapd <- data.frame(cellID = rownames(mapd), mapd)

cnr <- addQC(cnr, df = mapd, by = "cellID", sort = FALSE)

}
