% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ampCol.R
\docType{data}
\name{ampCol}
\alias{ampCol}
\title{ampCol -- ComplexHeatmap scale}
\format{
An object of class \code{function} of length 1.
}
\source{
\url{https://github.com/SingerLab/GAC"}
}
\usage{
data(ampCol)
}
\description{
ampcol is a ComplexHeatmap scale to use when plotting X data or gene bin data

Unlike segCol, the color pallete in ampCol is YlRd from RColorBrewer
}
\references{
\url{https://jokergoo.github.io/ComplexHeatmap-reference/book/}
}
\keyword{datasets}
