% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_diversity.R
\name{avg_num_alleles_per_locus}
\alias{avg_num_alleles_per_locus}
\title{Estimate the estimated number of alleles (copies) per locus}
\usage{
avg_num_alleles_per_locus(cnr, exclude.chr = NULL)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{exclude.chr}{chromosomes to exclude, default X, Y}
}
\value{
Estimates the number of alleles (copies per locus), across the population.

Function is `experimental` and requires validation to asses it's value
}
\description{
Function is `experimental` and requires validation to asses it's value
}
\examples{
data(cnr)

avg_num_alleles_per_locus(cnr)

avg_num_alleles_per_locus(cnr, exclude.chr = c("X", "Y"))

}
