% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary.cnr.R
\name{binary.cnr}
\alias{binary.cnr}
\title{build binary CNR from integer copy number data}
\usage{
binary.cnr(cnr, base.ploidy = 2)
}
\arguments{
\item{cnr}{a copy number matrix to convert to an incidence matrix}

\item{base.ploidy}{expected cell ploidy, e.g. 2N = 2, 4N = 4.}
}
\value{
Returns an incidence matrix for X as part of the cnr object

\itemize{
  \item Z incidence matrix from X, copy number != 2 is 1, all 2 are 0
}
}
\description{
This function builds a binary, incidence, matrix from the cnr$X.  It was designed as a helper function to generate the input for infSCITE.
}
\details{
Because infSCITE was developed for mutation data, this function creates a precense/absence matrix of the data

By default, anything not diploid is 1.
}
\examples{

data(cnr)

Z <- binary.cnr(cnr)
 
}
