% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary.cnr.R
\name{binary.X}
\alias{binary.X}
\title{build binary matrix from integer copy number data}
\usage{
binary.X(X)
}
\arguments{
\item{X}{a copy number matrix to convert to an incidence matrix}
}
\value{
Returns an incidence matrix for X, when X != 2
}
\description{
This function builds a binary, incidence, matrix from the cnr$X.  It was designed as a helper function to generate the input for infSCITE.
}
\details{
Because infSCITE was developed for mutation data, this function creates a precense/absence matrix of the data

By default, anything not diploid is 1.
}
\examples{

data(cnr)

Z <- binary.X(cnr$genes[, c("CDK4", "MDM2")])


}
