% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryDDRC.R
\name{binaryDDRC}
\alias{binaryDDRC}
\title{binaryDDRC}
\usage{
binaryDDRC(ddrc, base.ploidy = 2)
}
\arguments{
\item{ddrc}{a DDRC.df or DDRC.g matrix belonging to a cnr object}

\item{base.ploidy}{expected cell ploidy, e.g. 2N = 2, 4N = 4.}
}
\value{
returns a binary/incidence matrix for a DDRC object.  A value of 1 is
 returned for all values not equal to the cell ploidy.  Currently being
used to create an infScite table using clones.
}
\description{
binaryDDRC
}
\examples{

data(cnr)

noisy.cells <- cnr$qc$cellID[cnr$qc$qc.status == "FAIL"]

## reduced pipeline to genrate DDRC clone profiles
cnr <- excludeCells(cnr, excl = noisy.cells)
cnr <- phyloCNR(root.cell = "cell0")
cnr <- setBrayClusters()
cnr <- consensusClusterCNR(iters = 20, maxK = 40)
cnr <- doKSpectral()
cnr <- setKcc()
cnr <- cluster_heterogeneity(by = "category1",
          cluster_column = "ConsensusC")
cnr <- get_cluster_profiles()

binary.ddrc <- binaryDDRC(cnr$DDRC.df)
head(binary.ddrc)

}
