% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callCN.R
\name{callCN}
\alias{callCN}
\title{Copy number call categories from ratio data}
\usage{
callCN(
  cnr,
  cbioportal = FALSE,
  hyper.amplification = TRUE,
  deletion.threshold = 0,
  loss.threshold = 1,
  base.ploidy = 2,
  gains.thresholds = c(3, 5),
  amplification.thresholds = c(6, 20)
)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{cbioportal}{logical, weather to use the cBioPortal copy number
call categories of -2, -1, 0, 1, 2 to denote deletion, loss, neutral,
gain, amplificiation, respectively. Default is FALSE.  If TRUE,
hyper amplification calls will be joined with 2}

\item{hyper.amplification}{logical, weather to call hyper amplifications where
the number of copies exceeds the upper bound amplification threshold defaults
to TRUE, and uses 20 as the default. Parameter is ignored if cBioPortal
notation is TRUE}

\item{deletion.threshold}{integer to use as locus deletion, default = 0}

\item{loss.threshold}{number of copies to use as copy number loss, default = 1}

\item{base.ploidy}{integer to denote base ploidy to use as Neutral}

\item{gains.thresholds}{number of copies to use as copy number gains,
default to a ower bound of 3 (+1 copy from 2N) and upper bound of 5 (+3 copies)}

\item{amplification.thresholds}{number of copies to use as copy number
amplification defaults to a lower bound of 6 (+4 copies) and upper bound of
20 (+18 copies)}
}
\value{
With default parameters, the function returns a matrix with copy number
calls in the form of D, L, N, G, A and A2 from an genotype X matrix.  

| CN    |  Call  | cBioPortal Notation |
|------:|: ---- :|: ----------------- :|
|     0 |   D    | -2                  |
|     1 |   L    | -1                  |
|     2 |   N    |  0                  |
|   3-5 |   G    |  1                  |
|  6-20 |   A    |  2                  |
|   >20 |   A2   |  2                  |
}
\description{
Copy number call categories from ratio data
}
\examples{
data(cnr)

Xc <- callCN(cnr)
Xc <- callCN(cnr, hyper.amplification = FALSE)

Xp <- callCN(cnr, cbioportal = TRUE)

}
