% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callCN.R
\name{callDNA_CN}
\alias{callDNA_CN}
\title{call copy numbers out of CNR data}
\usage{
callDNA_CN(
  cnr,
  cbioportal = FALSE,
  deletion.threshold = -0.5,
  loss.threshold = -0.1,
  gains.threshold = 0.1,
  amplification.threshold = 1
)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{cbioportal}{logical, weather to use the cBioPortal copy number
call categories of -2, -1, 0, 1, 2 to denote deletion, loss, neutral,
gain, amplificiation, respectively. Default is FALSE.  If TRUE,
hyper amplification calls will be joined with 2}

\item{deletion.threshold}{integer to use as locus deletion, default -0.5}

\item{loss.threshold}{copy number ratio to use as copy number loss,
default -0.1}

\item{gains.threshold}{copy number ratio to use as copy number gains,
default 0.1}

\item{amplification.threshold}{copy number ratio to use as copy number
amplification default 1}
}
\value{
returns a matrix with copy number calls in the form of D, L, N, G, A1 and A2
from an X matrix

| CN        |  Call  | cBioPortal Notation |
|----------:|: ---- :|: ----------------- :|
|     -0.50 |   D    |  -2                 |
|     -0.10 |   L    |  -1                 |
|-0.1 - 0.1 |   N    |  0                  |
| 0.1 - 1.0 |   G    |  1                  |
|      >1.0 |   A    |  2                  |
}
\description{
call copy numbers out of CNR data
}
\examples{
data(cnr)

Xc <- callDNA_CN(cnr)

Xp <- callDNA_CN(cnr, cbioportal = TRUE)

}
