% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccClustCol.R
\docType{data}
\name{ccClustCol}
\alias{ccClustCol}
\title{ccClustCol -- ComplexHeatmap color scale}
\format{
An object of class \code{function} of length 1.
}
\source{
\url{https://github.com/SingerLab/gac/"}
}
\usage{
data(chromInfo)
}
\description{
ccClustCol is a ComplexHeatmap color scale to use when plotting the
consensus matrix
}
\keyword{datasets}
