% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_chromosome_annotations.R
\name{chr_breaks}
\alias{chr_breaks}
\title{Estimate chromosome end locations along a continuous genome}
\usage{
chr_breaks(cnr, genes = FALSE)
}
\arguments{
\item{cnr}{a cnr}

\item{genes}{weather to use gene data instead of bin, default is FALSE}
}
\value{
A named vector of chromosome breaks locations in the data.
Useful when adding lines to separate chromosomes,  or
a background when highlighting a chromosome
}
\description{
Estimate chromosome end locations along a continuous genome
}
