% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_chromosome_annotations.R
\name{chr_colors}
\alias{chr_colors}
\title{Generate chromosome sidebar colors}
\usage{
chr_colors(cnr, col = c("#404040", "#BABABA"), bin = TRUE)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{col}{alternating chromosome colors, default is c("#404040", "#BABABA")}

\item{bin}{weather to use bin or gene data, default is true}
}
\value{
A named vector of default chromosome colors
}
\description{
Generate chromosome sidebar colors
}
