% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromInfo.R
\docType{data}
\name{chromInfo}
\alias{chromInfo}
\title{Bin chromosome-basepair coordinates}
\format{
A data frame with chromosome and position information
\itemize{

 \item chr, chromosome

 \item end, base pair coordinate

}
}
\source{
\url{https://github.com/SingerLab/GAC}
}
\usage{
data(chromInfo)
}
\description{
A dataset containing the chromsome and end coordiantes for all the bins of the single-cell copy number matrix `copynumbers`
This table is cross referenced to create the `genes` table, an interpolated table to link the bins to genes
}
\keyword{datasets}
