% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_heterogeneity.R
\name{cluster_heterogeneity}
\alias{cluster_heterogeneity}
\title{cluster heterogeneity across the sample}
\usage{
cluster_heterogeneity(cnr, by = NULL)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{by}{a categorical variable used to stratify the cell population.  If
NULL i.e. no stratification, the representation will be done overall}
}
\description{
cluster heterogeneity across the sample
}
\examples{

## cluster table, no subsamples
cc <- table(sample(paste0("C", 1:12), size = 1000, replace = TRUE))

( cc <- cluster_representation(cc) )

cc <- data.frame(cluster = sample(paste0("C", 1:12), size = 1000, replace = TRUE),
            sample = sample(paste0("S", 1:6), size = 1000, replace = TRUE))

cc <- table(cc[, c("cluster", "sample")])
cc[cc <= 10] <- 0

( cc <- cluster_representation(cc) )

}
