% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_cnr.R
\name{cnr_data_type}
\alias{cnr_data_type}
\title{data type in CNR}
\usage{
cnr_data_type(cnr, detailed = FALSE)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{detailed}{logical, if TRUE print data type with data range, default FALSE}
}
\description{
data type in CNR
}
\examples{
data(cnr)
cnr_data_type(cnr)

cnr_data_type(cnr, detailed = TRUE)

}
