% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_representation.R
\name{consensus_representation}
\alias{consensus_representation}
\title{cluster representation across the sample}
\usage{
consensus_representation()
}
\arguments{
\item{cnr}{a cnr bundle}
}
\description{
cluster representation across the sample
}
