% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensusClusterCNR.R
\name{consensusClusterCNR}
\alias{consensusClusterCNR}
\title{run consensus clustering}
\usage{
consensusClusterCNR(
  cnr,
  maxK = 8,
  reps = 150,
  title = "cnr_ccp",
  innerLinkage = "ward.D2",
  finalLinkage = "ward.D2",
  seed = 2020.0314,
  ...
)
}
\arguments{
\item{cnr}{a cnr object}

\item{maxK}{maximum K, end of K sequence}

\item{reps}{number of times the clustering runs}

\item{title}{title for the consensus cluster map}

\item{innerLinkage}{inner linkage, see ConsensusClusterPlus}

\item{finalLinkage}{final linkage, see ConsensusClusterPlus}

\item{seed}{seed of analysis}

\item{...}{additional parameters for ConsensusClusterPlus e.g. plot = "png"}
}
\value{
A cnr object that contains the output of ConsensusClusterPlus for the
 specified Ks.

\itemize{
  \item ccp ConsensusClusterPlus object
}
}
\description{
Function runs consensus clustering for a pre-determined sequence of K.
}
\examples{
data(cnr)

cnr <- phyloCNR(cnr)

cnr <- consensusClusterCNR(cnr, maxK = 6)

}
