% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_chromosome_annotations.R
\name{create_chromosome_annotation}
\alias{create_chromosome_annotation}
\title{Create chromosome annotations for custom heatmaps}
\usage{
create_chromosome_annotation(cnr, side = "left", ...)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{side}{side of the annotation. Options are "left" default,
"right", "top", and "bottom"}

\item{...}{additional prameters passed to \code{\link[ComplexHeatmap]{HeatmapAnnotation}}}
}
\value{
A HeatmapAnnotation containing a chromosome map indicates chromosome
boundaries on heatmaps
}
\description{
Create chromosome annotations for custom heatmaps
}
\examples{
data(cnr)

chrAnnoLeft <- create_chromosome_annotation(cnr)
chrAnnoTop <- create_chromosome_annotation(cnr, side = "top")

}
