% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_chromAnno.R
\name{create_chromosome_annotation_right}
\alias{create_chromosome_annotation_right}
\title{create chromosome annotations for custom heatmaps}
\usage{
create_chromosome_annotation_right(cnr, ...)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{...}{additional prameters passed to HeatmapAnnotation}
}
\description{
create chromosome annotations for custom heatmaps
}
