% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depmap_annotate.R
\name{depmap_annotate}
\alias{depmap_annotate}
\title{annotate essential genes w/DepMap}
\usage{
depmap_annotate(cnr, depMeans, cutoff = -0.5)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{depMeans}{an object with the dependency score means for the cells
of interest}

\item{cutoff}{cutoff of D2 score to consider essential gene or
non-essentail gene}
}
\value{
Function returns the gene.index with depmap annotation colums;
 both depMeans, and essential (essentian/non-essential)
}
\description{
Takes a vector of depmap genes, and annotates the genes as esseintail
or not essential given a cutoff.  For details and downloads please visit
https://depmap.org/portal/
}
\examples{

\dontrun{
d2_map <- read.csv("inst/extdata/D2_combined_gene_dep_scores.csv.gz",
 header = TRUE, row.names = 1)

## select cells to use
keep <- grep("BONE", names(depmap))

d2_bone <- DepScores(depmap = d2_map, keep = keep)

depMeans <- rowMeans(d2_bone[,keep])

depmap_annotate(cnr, depMeans = depMeans)

}

}
