% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DepScores.R
\name{DepScores}
\alias{DepScores}
\title{Read DepMap D2_combined file}
\usage{
DepScores(depmap, keep)
}
\arguments{
\item{depmap}{Genetic Dependency Combined RNAi; Depmap D2_conbined object  (see: https://depmap.org/portal/)}

\item{keep}{which cell lines you want to keep}
}
\value{
Returns the depscores for a set of cells of interest
}
\description{
This function is file specific to the Depmap project D2_combined file
}
\details{
It will read, and process the row.names to preserve the hgnc.symbols and entrezID
}
\examples{

\dontrun{

d2_map <- read.csv("inst/extdata/D2_combined_gene_dep_scores.csv", header = TRUE, row.names = 1)

## select cells to use
keep <- grep("BONE", names(depmap))

d2_bone <- DepScores(depmap = d2_map, keep = keep)


depMeans <- rowMeans(d2_bone[,keep])

}

}
