% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distCNR.R
\name{distCNR}
\alias{distCNR}
\title{calculating cell-to-cell distances for clustering}
\usage{
distCNR(cnr, method = "bray", ...)
}
\arguments{
\item{cnr}{a cnr object}

\item{method}{method for calculating distances, defaults to Bray-Curtis
dissimilarity}

\item{...}{other parameters passed to vegan::vegdist}
}
\value{
Returns a cell-to-cell distance matrix of class `dist`
}
\description{
calculating cell-to-cell distances for clustering
}
\keyword{internal}
