% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doKSpectral.R
\name{doKSpectral}
\alias{doKSpectral}
\title{Run Specral K to select optimum and maximum K from Consensus Clustering}
\usage{
doKSpectral(cnr, ...)
}
\arguments{
\item{cnr}{a cnr bundle, with `ccp`}

\item{...}{additional parameters passed}
}
\value{
doKSpectral returns two matrices

kStats a dataframe containing number of stable K clusters (spectral K) for every kCC (K parameter), and the maximum delta Lamdba.

eigenVals a data frame containing the eigen values for each kCC

Lastly, it returns optK, a vector containing the recomended kCC and the value of `k` i.e. (spectral K). These are the kCC for the max number of k and max dLambdaMax
}
\description{
Run Specral K to select optimum and maximum K from Consensus Clustering
}
\references{
Philip A. Knight (2008) The Sinkhorn–Knopp Algorithm: Convergence and Applications. SIAM Journal on Matrix Analysis and Applications 30(1), 261-275. doi: 10.1137/060659624
}
\author{
Nick Socci <soccin@mskcc.org>
}
