% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kSpectral.R
\name{eps}
\alias{eps}
\title{eps : Spacing of floating point numbers}
\usage{
eps(N)
}
\arguments{
\item{N}{number of rows of S
#'}
}
\description{
eps : Spacing of floating point numbers
}
\references{
Philip A. Knight (2008) The Sinkhorn–Knopp Algorithm: Convergence and Applications. SIAM Journal on Matrix Analysis and Applications 30(1), 261-275. doi: 10.1137/060659624
}
\author{
Nick Socci <soccin@mskcc.org>
}
