% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_reg_histology.R
\name{estimate_joint_effects_bin}
\alias{estimate_joint_effects_bin}
\title{Joint CNA effect estimation using GLMNet at the bin level}
\usage{
estimate_joint_effects_bin(cnr, trait, pheno0, pheno1, exclude.cluster = "HC")
}
\arguments{
\item{cnr}{a cnr bundle}

\item{trait}{character, name of the trait of interest to analyze.  Must
be a column in the phenotype matrix (Y). e.g. "binary1"}

\item{pheno0}{character, phenotype(s) to use as baseline, e.g. "0"}

\item{pheno1}{character, phenotype(s) to use as alternate, e.g. "1"}

\item{exclude.cluster}{character, list of clusters to exclude, e.g.
hypersegmented, Stroma, etc. Default "HC"}
}
\value{
A CNR object containing joint CNA effect estimates using GLMNet at the
`BIN` level.  Hence, this function runs the analysis on the `X` matrix.

Effects are estimated at two Lambda; min, 1se.  Lambda values are appended
as attributes to the chromInfo, and gene.index, for the bin, respectively.
}
\description{
Internal function to \link{estimate_joint_effects}
}
\examples{
data(cnr)

cnr <- estimate_joint_effects(cnr, trait = "binary1",
   pheno0 = 0, pheno1 = 1)

}
\keyword{internal}
