% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excludeCells.R
\name{excludeCells}
\alias{excludeCells}
\title{exclude cells from CNR bundle}
\usage{
excludeCells(cnr, excl)
}
\arguments{
\item{cnr}{the cnr bundle}

\item{excl}{a string vector of cellID to be removed}
}
\value{
Returns a cnr object after removal of cells
}
\description{
Both keepCells and excludeCells perform similar functions; keep cells with positive selection, and excludeCells with negative selection
}
\examples{

data(cnr)

cnr2 <- excludeCells(cnr, excl = cnr$Y$cellID[cnr$qc$ReadsKept < 800000])

## not run: similarly:


cnr3 <- keepCells(cnr, keep = cnr$Y$cellID[cnr$qc$ReadsKept >= 800000])

all.equal(cnr2, cnr3)


}
