% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand2genes.R
\name{expand2genes}
\alias{expand2genes}
\title{Expands the bins to genes}
\usage{
expand2genes(X, gene.index, bin.id = "bin.id", gene.id = "hgnc.symbol")
}
\arguments{
\item{X}{the X bins data.frame}

\item{gene.index}{gene index with bin.id and hgnc.symbol}

\item{bin.id}{name of the bin.id column (the bin.id is the row in the bins data.frame)}

\item{gene.id}{string noting which gene ID you want to use e.g. hgnc.symbol, ensembl_gene_id}
}
\description{
This function interpolates the bins into an interpretable gene copy number matrix. All genes within the bin coordinates get assigned the same value. Some genes sit between bins without copy number data and will be marked as NA
}
\details{
The gene copy number matrix is core to the cnr object.
}
