% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportCNR.R
\name{export_cnr}
\alias{export_cnr}
\title{export CNR to flat text files}
\usage{
export_cnr(cnr, outdir = ".", ...)
}
\arguments{
\item{cnr}{a cnr}

\item{outdir}{path to output directory}

\item{...}{arguments passed to write.table and ape::write.tree}
}
\value{
Writes X, Y, genes, qc, gene.index, and chromInfo to `outdir`.  If
`phylo` exists, a `.newick` format tree is also exported. Furthermore,
if `DDRC` profiles are also present, files for these are also written
to the specified directory.
}
\description{
export CNR to flat text files
}
\examples{
\dontrun{
data(cnr)

export_cnr(cnr, outdir = "cnr_out/")
}

}
