% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_reg_histology.R
\name{export_pval_igv}
\alias{export_pval_igv}
\title{Export the p-values for comparisons of interest to visualize in IGV}
\usage{
export_pval_igv(
  cnr,
  pval.column,
  outdir = ".",
  outfile.prefix,
  extension = ".logistic",
  na.value = "1",
  print.output = FALSE
)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{pval.column}{column to export}

\item{outdir}{output directory, default is "."}

\item{outfile.prefix}{basename of file}

\item{extension}{extesion compatible with IGV. Any of ".linear",
".logistic", ".assoc", ".qassoc", ".gwas".  Default ".logistic"}

\item{na.value}{value to give to NA, default 1 (i.e. p-value = 1)}

\item{print.output}{print output on console as an object list with
bin.pvalues and gene.pvalues}
}
\value{
Exports p-values from an analysis into two files, one containing "bin" level
p-values, and another containing "gene" level p-values in IGV's GWAS format.

Bin-level p-values are expanded to all genes within each bin, such that
the user can visualize the bin blocks with added resolution.

By default, any p-values with  NA are converted to 1.  These can arise in
monomorphic regions of the genome and don't have an association to
the phenotype.
}
\description{
Export the p-values for comparisons of interest to visualize in IGV
}
\references{
https://software.broadinstitute.org/software/igv/GWAS
}
