% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgaCol.R
\docType{data}
\name{fgaCol}
\alias{fgaCol}
\title{fgaCol -- ComplexHeatmap Legend}
\format{
An object of class \code{function} of length 1.
}
\source{
\url{https://github.com/SingerLab/GAC"}
}
\usage{
data(chromInfo)
}
\description{
fgaCol is a ComplexHeatmap scale to use when plotting FGA data.
}
\keyword{datasets}
