% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scClustCNR.R
\name{fisher_tree}
\alias{fisher_tree}
\title{implementation of fisher_dist, hc, and tree}
\source{
\url{https://github.com/KrasnitzLab/SCclust}
}
\usage{
fisher_tree(cnr, hc.method = "average", tree.method = "average", ...)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{hc.method}{method on hcclust_tree, defaults to average}

\item{tree.method}{method on tree_py}

\item{...}{additional parameters pased to SCclust::sim_fisher_wrapper}
}
\description{
implementation of fisher_dist, hc, and tree
}
\keyword{internal}
