% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.aggregate.R
\name{gene.aggregate}
\alias{gene.aggregate}
\title{combine genes w/equal frequency to be only once}
\usage{
gene.aggregate(Z)
}
\arguments{
\item{Z}{binary incidence matrix or G genotype matrix.  Internally, both Z or G will be turned into a Z matrics for frequency calculations.}
}
\value{
returns a de-duplicated incidence or ternary matrix with rownams naving the multipe duplicated rows
}
\description{
This is a helper function to process the binary or ternary matrix.  Because gene data is interpolated from the bin data, linked loci within a bin will be duplicated.  Duplicate frequencies create infinite combinations in the trees in infSCITE.
}
\details{
This function checks if two rows have are identical and merges them into one. It also creates unique rownames to know what was de-duplicated
}
\examples{

data(cnr)

## for binary data
Z <- binary.cnr(cnr$genes[, c("CDK4", "MDM2", "HMGA2")])

aggrMat.Z <- gene.aggregate(Z = Z)

## for ternary data
G <- ternary.cnr(cnr$genes[, c("CDK4", "MDM2", "HMGA2")])

aggrMat.G <- gene.aggregate(Z = G)

}
