% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_vdj.R
\name{genotype_vdj}
\alias{genotype_vdj}
\title{Genotype cells at the VDJ recombination sites}
\usage{
genotype_vdj(cnr)
}
\arguments{
\item{cnr}{a cnr bundle.  gene.index MUST include a column named `gene.type`
with pattern "TR.*gene" for T-cell receptor genes, and "IG.*gene" for B-cell
receptor genes.  These are the defaults for gene.type in bioconductor::biomaRt}
}
\value{
Returns a CNR object with `cell.type` column containing assigments for
 `T-cell`, `B-cell`, and `vdj.unspecified`.  The function will only annotate
 cells with deletions on a `TR` or `IG` gene present in the gene.index table.
 All other annotation will be left intact.  If `cell.type` column is not
 present one will be created with `NA`.

An additional cnr[["vdj.cells"]] vector is also returned containing the cell
 type of only VDJ cells.
}
\description{
Genotype cells at the VDJ recombination sites
}
\examples{
data(cnr)

cnr <- genotype_vdj(cnr)

}
