% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_frequencies.R
\name{get_alt_counts}
\alias{get_alt_counts}
\title{alteration counts (both Amp + Del)}
\usage{
get_alt_counts(X, bulk, ...)
}
\arguments{
\item{X}{a matrix of integer copy number `cnr$X`}

\item{bulk}{logical, weather the cnr is from bulk DNA, default is FALSE}

\item{...}{additional parameters passed to get_amp_counts, get_del_counts, or
get_alt_counts}
}
\description{
alteration counts (both Amp + Del)
}
\keyword{internal}
