% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_frequencies.R
\name{get_alteration_frequenciesCNR}
\alias{get_alteration_frequenciesCNR}
\title{estimate bin and gene alteration frequencies}
\usage{
get_alteration_frequenciesCNR(cnr, symbol = "hgnc.symbol", ...)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{symbol}{gene symbol used on cnr[["genes"]] as column names, e.g. HGNC symbol or ensembl ID.
defaults to hgnc.symbol on package gene.index}

\item{...}{other parameters passed to merge}
}
\value{
Calls on get_bin_frequencies and get_gene_frequencies, respectively, to estimate amplification frequencies,
deletion frequencies, and the total of both alterations (amp + del combined) at every bin, and gene.

It then merges the bin frequencies to the chromInfo, and gene frequencies to the gene.index.

Currently only tested on integer copy number, where amplifications are hard-coded as three (3) or more copies,
and deletions as one (1) or zero (0).
}
\description{
Wrapper function to calculate alteration frequencies (amplifications, deletions, and both combined) at
the bin level, and gene level.
}
\examples{

data(cnr)

head(cnr$gene.index)
head(cnr$chromInfo)

cnr <- get_alteration_frequenciesCNR(cnr)

head(cnr$chromInfo)
head(cnr$gene.index)

}
