% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_frequencies.R
\name{get_amp_frequencies}
\alias{get_amp_frequencies}
\title{amplification gene frequency}
\usage{
get_amp_frequencies(X, bulk)
}
\arguments{
\item{X}{a matrix of integer copy number `cnr$X`}

\item{bulk}{logical, weather the cnr is from bulk DNA, default is FALSE}
}
\description{
amplification gene frequency
}
