% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cluster_profiles.R
\name{get_cluster_profiles}
\alias{get_cluster_profiles}
\title{Build a consensus copy number profile for each cluster or clone}
\usage{
get_cluster_profiles(cnr, minimum_cells = 3)
}
\arguments{
\item{cnr}{a cnr bundle containing the output of setCluster}

\item{minimum_cells}{minimum number of cells in a cluster, must be greater than 3
to estimate a median}
}
\value{
Returns a vector of clusters with miminum `n` cells (default 3), and two matrices.
 One matrix `DDRC.df` with the median consensus profile for the cluster using
 bins, the other `DDRC.g` which composes the median consensus profile for each
 gene.
}
\description{
Build a consensus copy number profile for each cluster or clone
}
