% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_frequencies.R
\name{get_del_counts}
\alias{get_del_counts}
\title{deletion counts}
\usage{
get_del_counts(X, bulk, base.ploidy = 2)
}
\arguments{
\item{X}{a matrix of integer copy number `cnr$X`}

\item{bulk}{logical, weather the cnr is from bulk DNA, default is FALSE}

\item{base.ploidy}{tumor expected ploidy}
}
\description{
deletion counts
}
\keyword{internal}
