% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_gene_details.R
\name{get_gene_details}
\alias{get_gene_details}
\title{Pull gene details for a genomic region}
\usage{
get_gene_details(cnr, seqnames = 12, start = 69200804, end = 69246466)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{seqnames}{a chromosome name, must match 'cnr$gene.index$seqnames'}

\item{start}{region start}

\item{end}{region end}
}
\value{
Returns the subset of the `gene.index` table for the genomic region.
}
\description{
This function subsets the gene index for a genomic region of interest.
}
\examples{

data(cnr)
coord.df <- data.frame(chr = 12,
                    start = 69200804,
                    end = 69246466)

get_gene_details(cnr, seqnames = coord.df$chr,
                  start = coord.df$start, end = coord.df$end)

coords.df <- data.frame(chr = c(1, 12), 
                     start = c(170120554, 69200804),
                     end =  c(172941951, 69246466))
do.call(rbind, apply(coords.df, 1, function(rr)
                     get_gene_details(cnr,
                                       seqnames = rr[1],
                                       start = rr[2],
                                       end = rr[3])))

}
