% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scClustCNR.R
\name{getSubclonesCNR}
\alias{getSubclonesCNR}
\title{implementation of finding clones and subclones from SCclust}
\source{
\url{https://github.com/KrasnitzLab/SCclust}
}
\usage{
getSubclonesCNR(cnr, ...)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{...}{additional parameters for find_subclones}
}
\description{
implementation of finding clones and subclones from SCclust
}
