% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisticGR.R
\name{gisticGR}
\alias{gisticGR}
\title{Gistic Genomic Ranges, get genes in gene.index that overlap with significant genomic ranges in GISTIC2 output}
\usage{
gisticGR(cnr, grTR, oncokb, depmeans, cutoff = -0.5)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{grTR}{gistic regions table, output from gisticRegions}
}
\description{
**Still in development**
** requires GenommicRanges
}
