% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisticRegions.R
\name{gisticRegions}
\alias{gisticRegions}
\title{Process GISTIC2 output // output only grTR
read in gistic leasions file
provide path to gistic directory w/ all_lesions.conf_**.txt}
\usage{
gisticRegions(gisticDir, conf = 80)
}
\arguments{
\item{gisticDir}{path to directory of gistic output}

\item{conf}{confidence used in gistic, default is 75, for single-cells 80 or highrer works a little better.  Also, requires increasing amp/del calling thresholds to log2(1/2) for deletions, and log2(3/2) for gains and amplifications}
}
\description{
**Still in development**
}
\examples{
grTR <- gisticRegions("path/to/gistic/", conf = 80)

}
