% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grch37.genes.5k.R
\docType{data}
\name{grch37.genes.5k}
\alias{grch37.genes.5k}
\title{Bin-to-gene index}
\format{
A data frame with chromosome and position information
\itemize{

  \item hgnc.symbol, HGNC gene symbol
  \item chrom, chromosome 
  \item start, gene start coordinate
  \item end, gene end coordinate
  \item width, width
  \item strand, gene strand
  \item ensembl_gene_id, ENSEMBL gene ID
  \item gene_biotype, gene type
  \item bin.id, bin ID

}
#' @docType data
}
\source{
\url{https://github.com/SingerLab/gac}
}
\usage{
data(grch37.genes.5k)
}
\description{
This table was constructed by mapping bin regions to gene coordinates.
 We found the simplest way to do this, was to pull the gencode annotation
with biomaRt[https://bioconductor.org/packages/release/bioc/vignettes/biomaRt/inst/doc/biomaRt.html]
and use findOverlap with
GenomicRanges[https://bioconductor.org/packages/release/bioc/vignettes/GenomicRanges/inst/doc/GenomicRangesIntroduction.html].
The resulting GRanges object was saved as a text file.
}
\details{
A table to interpolate bin copy number to individual genes.
}
\keyword{datasets}
