% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distCNR.R
\name{hclustCNR}
\alias{hclustCNR}
\title{heirarchical clustering}
\usage{
hclustCNR(cnr, method = "ward.D2", ...)
}
\arguments{
\item{cnr}{a cnr object}

\item{method}{method for heirarchical clustering, defaults to "ward.D2"}

\item{...}{other parameters passed to hclust}
}
\value{
Returns a heirarchical clustering object
}
\description{
heirarchical clustering
}
