% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hg19_cytoBand.R
\docType{data}
\name{hg19_cytoBand}
\alias{hg19_cytoBand}
\title{HG19 cytoband file from UCSC genome browser}
\format{
A matrix containg the cytoband coordinates

\itemize{
   \item bins, in rows from j..n
   \item cellIDs, in columns from i..n
}
}
\source{
\url{https://github.com/SingerLab/gac}
}
\usage{
data(hg19_cytoBand)
}
\description{
Human reference GRCh37/hg19 cytoband file from UCSC Genome Browser

File is used by SCclust to remove bins on the chromosome centromeres
}
\keyword{datasets}
