% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_reg_histology.R
\name{histo_logit_bin}
\alias{histo_logit_bin}
\title{perform logistic regresion for a binarized trait}
\usage{
histo_logit_bin(
  cnr,
  trait,
  pheno0,
  pheno1,
  exclude.cluster = "HC",
  family = "binomial",
  na.action = "na.exclude",
  ...
)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{trait}{character, name of the trait of interest to analyze.  Must
be a column in the phenotype matrix (Y). e.g. "binary1"}

\item{pheno0}{character, phenotype(s) to use as baseline, e.g. "0"}

\item{pheno1}{character, phenotype(s) to use as alternate, e.g. "1"}

\item{exclude.cluster}{character, list of clusters to exclude, e.g.
hypersegmented, Stroma, etc. Default "HC"}

\item{family}{character, description of the error distribution and
link function to be used in the model. See \link[stats]{glm} for details.
 Default "binomial"}

\item{na.action}{character, handling of NA. default is "na.exclude"}

\item{...}{additional arguments passed to glm}
}
\value{
a CNR object with results from a logistic regression analysis
 (family = "binomial") with effect estimates, and p-values attached to
 the chromInfo matrices.

Results columns are "Estimate", "Std.Error", "z.value", "p.value" ; with
 the phenotype comparison pre-apended as <pheno0>.vs.<pheno1>.lr.<value>.
 Using grade as an example  these would be:

0.vs.1.lr.Estimate
0.vs.1.lr.Std.Error
0.vs.1.lr.z.value
0.vs.1.lr.p.value
}
\description{
Internal to histo_logit for bin level analysis
}
\examples{

data(cnr)

cnr <- histo_logit(cnr, trait = "binary1",
   pheno0 = 0, pheno1 = 1)

cnr <- histo_logit(cnr, trait = "category1",
   pheno0 = "A", pheno1 = c("B", "C"))

cnr <- histo_logit(cnr, trait = "category1",
   pheno0 = c("X", "Y"), pheno1 = "Z")

}
