% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_reg_histology.R
\name{histo_logit_cov}
\alias{histo_logit_cov}
\title{perform logistic regression for a binary/binarized trait with a covaritate}
\usage{
histo_logit_cov(
  cnr,
  trait,
  pheno0,
  pheno1,
  covar,
  exclude.cluster = "HC",
  family = "binomial",
  na.action = "na.exclude",
  ...
)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{trait}{character, name of the trait of interest to analyze.  Must
be a column in the phenotype matrix (Y). e.g. "binary1"}

\item{pheno0}{character, phenotype(s) to use as baseline, e.g. "0"}

\item{pheno1}{character, phenotype(s) to use as alternate, e.g. "1"}

\item{covar}{character, model covariates to include in the model,
e.g. "category1"}

\item{exclude.cluster}{character, list of clusters to exclude, e.g.
hypersegmented, Stroma, etc. Default "HC"}

\item{family}{character, description of the error distribution and
link function to be used in the model. See \link[stats]{glm} for details.
Default "binomial"}

\item{na.action}{character, handling of NA. default is "na.exclude"}

\item{...}{additional arguments passed to glm}
}
\value{
a CNR object with results from a logistic regression analysis
(family = "binomial") with effect estimates, and p-values attached to
the chromInfo and gene.index matrices.

Results columns are "Estimate", "Std.Error", "z.value", "p.value" ; with
 the phenotype comparison pre-apended as
 <pheno0>.vs.<pheno1>.cv<covar>.lr.<value>.
 Using grade as an example  these would be:

0.vs.1..lr.Estimate
0.vs.1.quantitative1.lr.Std.Error
0.vs.1.quantitative1.lr.z.value
0.vs.1.quantitative1.lr.p.value
}
\description{
These functions were developed for the genetic analysis of cancer subtypes,
where "histology" is considered the phenotype, and copy number the genotype.
The association is carried out using logistic regresion by `glm` using
`family = "binary"`, as this recreates the standard model of quantitative
genetics P = G + E.  Alternate arguments of `family` have no been tested.
}
\examples{

pdata(cnr)

cnr <- histo_logit_cov(cnr, trait = "binary1",
   pheno0 = 0, pheno1 = 1, covar = "category1")

cnr <- histo_logit_cov(cnr, trait = "category1",
   pheno0 = "A", pheno1 = c("B", "C"), covar = "category2")

cnr <- histo_logit_cov(cnr, trait = "category1",
   pheno0 = c("X", "Y"), pheno1 = "Z", covar = "category1")

}
