% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keepCells.R
\name{keepCells}
\alias{keepCells}
\title{keep cells from CNR bundle}
\usage{
keepCells(cnr, keep)
}
\arguments{
\item{cnr}{the cnr bundle}

\item{keep}{a string vector of cellID to keep}
}
\value{
Returns a CNR bundle with only the cells we want to keep
}
\description{
Both keepCells and excludeCells perform similar functions; keep cells with positive selection, and excludeCells with negative selection
}
\examples{

data(cnr)

cnr2 <- excludeCells(cnr, excl = cnr$Y$cellID[cnr$qc$ReadsKept < 800000])

cnr3 <- keepCells(cnr, keep = cnr$Y$cellID[cnr$qc$ReadsKept >= 800000])

all.equal(cnr2, cnr3)

}
