% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kSpectral.R
\name{kSpectral}
\alias{kSpectral}
\title{kSpectral}
\usage{
kSpectral(S)
}
\arguments{
\item{S}{similarity matrix}
}
\description{
kSpectral
}
\references{
Philip A. Knight (2008) The Sinkhorn–Knopp Algorithm: Convergence and Applications. SIAM Journal on Matrix Analysis and Applications 30(1), 261-275. doi: 10.1137/060659624
}
\author{
ND Socci <soccin .at. mskcc dot org>
}
\keyword{internal}
