% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legAmp.R
\docType{data}
\name{legAmp}
\alias{legAmp}
\title{legAmp -- ComplexHeatmap Legend}
\format{
An object of class \code{Legends} of length 1.
}
\source{
\url{https://github.com/SingerLab/gac"}
}
\usage{
data(legAmp)
}
\description{
legAmp is a ComplexHeatmap Legend to use when plotting X data or gene bind ata

Unlike legSeg, the color palete here is `YrRd` from RColorBrewer. I use it
to highlight amplification genes
}
\keyword{datasets}
