% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legFGA.R
\docType{data}
\name{legFGA}
\alias{legFGA}
\title{legFGA -- ComplexHeatmap Legend}
\format{
An object of class \code{Legends} of length 1.
}
\source{
\url{https://github.com/SingerLab/gac"}
}
\usage{
data(legFGA)
}
\description{
legFGA is a ComplexHeatmap Legend to use when plotting FGA data.
}
\keyword{datasets}
