% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legLog2R.R
\docType{data}
\name{legLog2R}
\alias{legLog2R}
\title{legSeg -- ComplexHeatmap Legend}
\format{
An object of class \code{Legends} of length 1.
}
\source{
\url{https://github.com/SingerLab/gac"}
}
\usage{
data(chromInfo)
}
\description{
legLog2R is a ComplexHeatmap Legend to use when plotting X data in Log2 ratio form
e.g. bulk DNA data
}
\keyword{datasets}
