% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legPloidy.R
\docType{data}
\name{legPloidy}
\alias{legPloidy}
\title{legPloidy -- ComplexHeatmap Legend}
\format{
An object of class \code{Legends} of length 1.
}
\source{
\url{https://github.com/SingerLab/gac"}
}
\usage{
data(legPloidy)
}
\description{
legPloidy is a ComplexHeatmap Legend to use when plotting ploidy phenotypes.
}
\keyword{datasets}
