% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legSeg.R
\docType{data}
\name{legSeg}
\alias{legSeg}
\title{legSeg -- ComplexHeatmap Legend}
\format{
An object of class \code{Legends} of length 1.
}
\source{
\url{https://github.com/SingerLab/gac"}
}
\usage{
data(chromInfo)
}
\description{
legSeg is a ComplexHeatmap Legend to use when plotting X data
}
\keyword{datasets}
