% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_lookups.R
\name{list_genes}
\alias{list_genes}
\title{list out genes}
\usage{
list_genes(cnr, at, identifier = "hgnc.symbol")
}
\arguments{
\item{cnr}{a cnr bundle}

\item{at}{character, coordinate string e.g. "1:123456780:124567890", 
The \code{at} coordinates string is run through \code{convert_coord}, which
removes the preceeding `chr` and commas, and substitutes any dash `-`  with  `:` to
fit the format used here}

\item{identifier}{character, name of the gene identifier column to output,
default "hgnc.symbol"}
}
\value{
A vector containing a list of genes or other gene identifier.
}
\description{
list out genes
}
\examples{
data(cnr)

list_genes(cnr, at = "12:58000000:59000000")

}
