% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_lookups.R
\name{list_genes_in_region}
\alias{list_genes_in_region}
\title{finds genes within one chromosome interval}
\usage{
list_genes_in_region(cnr, at, show.columns = NULL, identifier = "hgnc.symbol")
}
\arguments{
\item{cnr}{a cnr bundle}

\item{at}{character, coordinate string}

\item{show.columns}{subset of columns from gene.index to output.
default is NULL, which shows all columns}

\item{identifier}{name of column with gene identifiers}
}
\value{
Returns the subset of the gene.index within the coordinates specified.
 By default all columns are returned, with the option to select a
 subset using \code{show.columns}.
}
\description{
finds genes within one chromosome interval
}
\examples{

data(cnr)

list_genes_in_region(cnr, at = "12:58,000,000-59,000,000")

list_genes_in_region(cnr, at = "12:58,000,000-59,000,000",
           show.columns = c("hgnc.symbol", "gene_biotype"))


}
