% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupCN.R
\name{lookupCN}
\alias{lookupCN}
\title{Lookup copy number data from bins}
\usage{
lookupCN(df, coord)
}
\arguments{
\item{df}{`data.frame` containing chrom, and chrompos}

\item{coord}{a `list` with elements named chr start end}
}
\description{
This function pulls the set of copy number data for a specific set of coordinates from a table, e.g. the CNR$chromInfo,
}
\examples{


data(cnr)
coords <- list(chr = 2, start = 550000, end = 600000)

lookupCN(df = cnr$chromInfo, coord = coords)

}
