% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowCol.R
\docType{data}
\name{lowCol}
\alias{lowCol}
\title{lowCol -- ComplexHeatmap color scale}
\format{
An object of class \code{function} of length 1.
}
\source{
\url{https://github.com/SingerLab/GAC"}
}
\usage{
data(lowCol)
}
\description{
lowCol is a ComplexHeatmap color scale to use when plotting X data. It's intended
use is for plotting DNA data, as this color scale is the familiar blue-to-red
colors used for bulk DNA copy number log ratios


#' @docType data
}
\keyword{datasets}
