% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapd.R
\name{mapd}
\alias{mapd}
\title{Median absolute pairwise deviance (MAPD)}
\usage{
mapd(Xi)
}
\arguments{
\item{Xi}{copy number data for one cell}
}
\value{
Returns the median deviance, median deviance sd, and median deviance cv
}
\description{
Median absolute pairwise deviance (MAPD)
}
\examples{
data(cnr)

mapd(cnr$X[,1])

}
