% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClusters.R
\name{maximum.percentage}
\alias{maximum.percentage}
\title{Max percent}
\usage{
maximum.percentage(mocp, low.boundary = 0.7, high.boundary = 0.9)
}
\arguments{
\item{mocp}{output from \code{\link{optClust}}}

\item{low.boundary}{lower percent of clones with multiple cells,
default is .70}

\item{high.boundary}{highest percent of clones with multiple cells,
default is .90}
}
\value{
The maximum tree height located at the maximum percent of clones with
 multiple cells between the low and high boundaries.
}
\description{
Max percent
}
