% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_lineage_tables.R
\name{merge_lineage_tables}
\alias{merge_lineage_tables}
\title{merge lineage tracing results from Corey's Maximum Entropy method}
\usage{
merge_lineage_tables(cnr, branch_table, lineage_table, adjacency_alterations)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{branch_table}{table named Branch_Table.csv}

\item{lineage_table}{table named Output_Table.csv}

\item{adjacency_alterations}{table named Tree_Edges.csv}
}
\description{
merge lineage tracing results from Corey's Maximum Entropy method
}
\examples{
\dontrun{
brc <- read.csv("data-raw/lineages/WD9048/Branch_Table.csv")
lin <- read.csv("data-raw/lineages/WD9048/Output_Table.csv")
adj <- read.csv("data-raw/lineages/WD9048/Tree_Edges.csv")

cnr <- merge_lineage_tables(WD9048a, brc, lin, adj)
}

}
