% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_chromosome_annotations.R
\name{mid_chr}
\alias{mid_chr}
\title{Estimate chromosome midpoint locations along a continuous genome}
\usage{
mid_chr(cnr, genes = FALSE)
}
\arguments{
\item{cnr}{a cnr}

\item{genes}{weather to use gene data instead of bin, default is FALSE}
}
\value{
A named vector of chromosome midpoints. Useful for adding tick
marks in figures.  Midpoint is not the centromere location.
}
\description{
Estimate chromosome midpoint locations along a continuous genome
}
