% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClusters.R
\name{minimum.intersect}
\alias{minimum.intersect}
\title{Minimum tree.height at the intersect of one-cell and multi-cell clusters}
\usage{
minimum.intersect(mocp)
}
\arguments{
\item{mocp}{output from \link{optClust}}
}
\description{
Minimum tree.height at the intersect of one-cell and multi-cell clusters
}
