% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryCNR.R
\name{ncells}
\alias{ncells}
\title{number of cells in cnr}
\usage{
ncells(cnr, detailed = FALSE, display = FALSE)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{detailed}{logical, detailed summary}

\item{display}{display cell ID's}
}
\description{
number of cells in cnr
}
