% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_cnr.R
\name{ngenes}
\alias{ngenes}
\title{number of genes in gene.index}
\usage{
ngenes(cnr, detailed = FALSE, display = FALSE)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{detailed}{logical, detailed summary, if TRUE prin number of genes by chromosome,
default FALSE}

\item{display}{logical, display gene.index}
}
\description{
number of genes in gene.index
}
\examples{
data(cnr)

ngenes(cnr)

ngenes(cnr, detailed = TRUE)


}
