% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_cnr.R
\name{npheno}
\alias{npheno}
\title{number of phenotypes}
\usage{
npheno(cnr, detailed = FALSE, display = FALSE)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{detailed}{detailed summary}

\item{display}{logical display example}
}
\description{
number of phenotypes
}
\examples{
data(cnr)

npheno(cnr)

npheno(cnr, detailed = TRUE)


}
