% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_cnr.R
\name{order_bins}
\alias{order_bins}
\title{order genome bins based on chromosome and starting position}
\usage{
order_bins(
  cnr,
  chromosome.order = c(1:22, "X", "Y", "MT"),
  chrom.column = "bin.chrom",
  start.column = "bin.start"
)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{chromosome.order}{order for chromosomes, default is 1:22, X, Y, and MT,
corresponding to the human genome}

\item{chrom.column}{column name for the bin chromosomes. default "bin.chrom"}

\item{start.column}{column name for bin start. default "bin.start"}
}
\value{
Function returns an chromInfo ordered by bin chromosomes and start coordinates
}
\description{
order genome bins based on chromosome and starting position
}
\examples{
data(cnr)

set.seed(2023)
shuffled.bins <- sample(1:nrow(cnr$chromInfo), size = nrow(cnr$chromInfo))
cnr$chromInfo <- cnr$chromInfo[shuffled.bins, ]
head(cnr$chromInfo)

cnrS <- order_bins(cnr)
head(cnrS$chromInfo)

}
