% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_cnr.R
\name{order_genes}
\alias{order_genes}
\title{order gene.index based on chromosome and starting coordinate}
\usage{
order_genes(
  cnr,
  chromosome.order = c(1:22, "X", "Y", "MT"),
  chrom.column = "chrom",
  start.column = "start"
)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{chromosome.order}{order for chromosomes, default is 1:22, X, Y, and MT,
corresponding to the human genome}

\item{chrom.column}{column name for the bin chromosomes. default "chrom"}

\item{start.column}{column name for bin start. default "start"}
}
\value{
Function returns an gene.index ordered by chromosomes and start
}
\description{
order gene.index based on chromosome and starting coordinate
}
\examples{
data(cnr)

set.seed(2023)
shuffled.genes <- sample(1:nrow(cnr$gene.index), size = nrow(cnr$gene.index))
cnr$gene.index <- cnr$gene.index[shuffled.genes, ]
head(cnr$gene.index)

cnrS <- order_genes(cnr)

head(cnrS$gene.index)

}
