% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parctan.R
\name{parctan}
\alias{parctan}
\title{Peter Andrew's arctan transformation for visualizing copy number}
\usage{
parctan(x)
}
\arguments{
\item{x}{copy number data}
}
\description{
The parctan transformation scales the data for visualization.  It zooms into low copy number, yet still see high valued copy number
}
\details{
See http://mumdex.com
}
\examples{
data(cnr)

pX <- parctan(cnr$X)

}
