% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathview_clones.R
\name{pathview_clones}
\alias{pathview_clones}
\title{plot pathview of specific clones}
\usage{
pathview_clones(
  cnr,
  clones,
  pathway.id = "hsa05200",
  species = "hsa",
  ploidy.norm = 2,
  gene.idtype = "SYMBOL",
  out.suffix = "clones",
  node.sum = "max",
  ...
)
}
\arguments{
\item{cnr}{a cnr bundle after running \code{\link{cluster_heterogeneity}}}

\item{clones}{a list of clones (names from DDRC.g)}

\item{pathway.id}{a pathview pathway ID use \code{link[pathview]{paths.hsa}}.
Default "hsa05200"}

\item{species}{use species from pathview. Default "hsa" for humans}

\item{ploidy.norm}{ploidy to normalize (e.g. if 2, then CN = 2 is now 0)}

\item{gene.idtype}{type of gene id in data matrix. GAC uses hgnc.symbol}

\item{out.suffix}{filename output suffix, will be appendd after the clones}

\item{node.sum}{how to summarise node data when gene has multiple mappings.
Possible values are: "sum","mean", "median", "max", "max.abs" and "random".
Passed to \code{\link[pathview]{node.map}}.  default is 'max.abs'.}

\item{...}{additional parameters passed to pathview}
}
\value{
Returns a normalized CN representation of a KEGG pathway.  By default, the scale
shows the copy number deviation from the expected ploidy of the tumor.

Output is written to the working directory as a PNG, among other files
}
\description{
this function is a wrapper to plot CNA detected in clones onto KEGG pathways.
Requires Bioconductor/Pathview to be installed
}
\examples{
data(cnr)
data(paths.hsa)

noisy.cells <- cnr$qc$cellID[cnr$qc$qc.status == "FAIL"]

## reduced pipeline to genrate DDRC clone profiles
cnr <- excludeCells(cnr, excl = noisy.cells)
cnr <- phyloCNR(root.cell = "cell0")
cnr <- setBrayClusters()
cnr <- consensusClusterCNR(iters = 20, maxK = 40)
cnr <- doKSpectral()
cnr <- setKcc()
cnr <- cluster_heterogeneity(by = "category1",
          cluster_column = "ConsensusC")
cnr <- get_cluster_profiles()

clone.1 <- colnames(cnr$DDRC.g)[1]

top.3.clones <- names(rev(sort(table(cnr$Y$final_cluster))))[1:3]

pathview_clones(cnr, clones = clone.1)
pathview_clones(cnr, clones = top.3.clones)

}
\references{
If you use this module, please cite:

Weijun Luo and Cory Brouwer. Pathview: an R/Bioconductor package for
pathway-based data integration and visu- alization. Bioinformatics,
29(14):1830-1831, 2013. doi: 10.1093/bioinformatics/btt285.
}
