% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_diversity.R
\name{percent_genome_amplified}
\alias{percent_genome_amplified}
\title{calculate the proportion of the genome amplified}
\usage{
percent_genome_amplified(cnr, by = "cell", amplification.threshold = 6, ...)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{by}{character, estimate percent genome loss at population, clone,
or cell levels. Options are "clone", and "cell".
Using "clone", and "cell", genome loss is performed using the clone profiles
(DDRC.df), and cell profiles (X), respectively. Default "cell"}

\item{amplification.threshold}{number of copies at which a locus is considered as an amplification.}

\item{...}{additional parameters to \code{\link{percent_genome_gain}}}
}
\description{
calculate the proportion of the genome amplified
}
\examples{

data(cnr)

noisy.cells <- cnr$qc$cellID[cnr$qc$qc.status == "FAIL"]

## reduced pipeline to genrate DDRC clone profiles
cnr <- excludeCells(cnr, excl = noisy.cells)
cnr <- phyloCNR(cnr, root.cell = "cell0")
cnr <- setBrayClusters(cnr)
cnr <- run_consensus_clustering(cnr, iters = 20, maxK = 40)
cnr <- doKSpectral(cnr)
cnr <- setKcc(cnr)
cnr <- cluster_heterogeneity(cnr, by = "category1",
          cluster_column = "ConsensusC")
cnr <- get_cluster_profiles(cnr)

percent_genome_amplified(cnr)

percent_genome_amplified(cnr, by = "cell")

percent_genome_amplified(cnr, by = "clone")

}
