% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno.R
\docType{data}
\name{pheno}
\alias{pheno}
\title{Single-cell Annotation data}
\format{
A data frame or matrix n.cell rows x y.phenotypes columns
\describe{
  \item{cellID}{single-cell ID}
  \item{trait1..n}{Number of reads kept after alignment}

  ...
}
}
\source{
\url{https://github.com/SingerLab/toSignac/}
}
\usage{
pheno
}
\description{
A dataset containing single-cell metadata, and phenotypic data
}
\keyword{datasets}
