% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_cnr.R
\name{phylo_ddrc}
\alias{phylo_ddrc}
\title{Calculating clone-to-clone distances, and estimating a phylogenetic tree}
\usage{
phylo_ddrc(
  cnr,
  root.clone = NULL,
  exclude.clones = NULL,
  dist.method = "bray",
  tree.method = "bal",
  spr = TRUE,
  ...
)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{root.clone}{a single clone to root the tree}

\item{exclude.clones}{vector of clones to exclude}

\item{dist.method}{method for calculating cell-to-cell distance
(see vegan::vegdist)}

\item{tree.method}{method for phylogenetic tree construction
either "bal", "ols", or "nj". default is "bal"}

\item{spr}{parameter specific to \link[ape]{fastme.bal}. Performs
Subtree Pruning and Regrafting.  Ignored in all other methods.
 Default is TRUE}

\item{...}{other parameters passed to root}
}
\value{
Creates a cell-to-cell distance matrix, runs heirarchical clustering.
 By default a phylogenetic tree is constructed using \link[ape]{fastme.bal},
alternative algorithms are \link[ape]{fastme.ols} or \link[ape]{nj}.

\itemize{
  \item DDRC.dist clone to clone Bray-Curtis dissimiarly 
  \item DDRC.phylo phylogenetic tree of class `phylo` from R/ape
}
}
\description{
Calculating clone-to-clone distances, and estimating a phylogenetic tree
}
\examples{

data(cnr)

cnr <- phylo_cnr(cnr)
cnr <- setBrayClusters(cnr)
cnr <- get_cluster_profiles(cnr, cluster.column = "BrayC")

## unrooted cell phylogenetic tree
cnr <- phylo_ddrc(cnr)

cnr$DDRC.phylo
plot(cnr$DDRC.phylo)


## rooted cell phylogenetic tree
cnr <- phylo_ddrc(cnr, root.clone = "C2")

cnr$DDRC.phylo
plot(cnr$DDRC.phylo)

## change method to ols
cnr <- phylo_ddrc(cnr, root.clone = "C2", tree.method = "ols")

cnr$DDRC.phylo
plot(cnr$DDRC.phylo)

}
\references{
Vincent Lefort,  Richard Desper,  Olivier Gascue. 2015. "FastME 2.0:
  A Comprehensive, Accurate, and Fast Distance-Based Phylogeny Inference
  Program". Molecular Biology and Evolution, Volume 32, Issue 10, October
  2015, Pages 2798–2800. <https://doi.org/10.1093/molbev/msv150?>

Paradis E, Schliep K (2019). “ape 5.0: an environment for modern
  phylogenetics and evolutionary analyses in R.” Bioinformatics,
  *35*, 526-528.  <https://doi.org/10.1093/bioinformatics/bty633>.

Oksanen J, Simpson G, Blanchet F, Kindt R, Legendre P, Minchin P,
  O'Hara R, Solymos P, Stevens M, Szoecs E, Wagner H, Barbour M,
  Bedward M, Bolker B, Borcard D, Carvalho G, Chirico M, De Caceres M,
  Durand S, Evangelista H, FitzJohn R, Friendly M, Furneaux B, Hannigan
  G, Hill M, Lahti L, McGlinn D, Ouellette M, Ribeiro Cunha E, Smith T,
  Stier A, Ter Braak C, Weedon J (2022). _vegan: Community Ecology
  Package_. R package version 2.6-4,
  <https://CRAN.R-project.org/package=vegan>.
}
