% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloCNR.R
\name{phyloDDRC}
\alias{phyloDDRC}
\title{Calculating clone-to-clone distances, and estimating a phylogenetic tree}
\usage{
phyloDDRC(
  cnr,
  root.clone = NULL,
  exclude.clones = NULL,
  dist.method = "bray",
  hclust.method = "ward.D2",
  ...
)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{root.clone}{a single clone to root the tree}

\item{exclude.clones}{vector of clones to exclude}

\item{dist.method}{method for calculating cell-to-cell distance
(see vegan::vegdist)}

\item{hclust.method}{method for heirarchical clustering (see hclust)}

\item{...}{other parameters passed to vegan}
}
\value{
Creates a cell-to-cell distance matrix, runs heirarchical clustering,
 and converts the `hclust` object to an `ape` class `phylo` object.

\itemize{
  \item cdb cell to cell Bray-Curtis dissimiarly 
  \item hcdb heirarchical clustering of distance matrix
  \item phylo ape class `phylo` object
}
}
\description{
Calculating clone-to-clone distances, and estimating a phylogenetic tree
}
