% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidyCol.R
\docType{data}
\name{ploidyCol}
\alias{ploidyCol}
\title{ploidyCol -- ComplexHeatmap Legend}
\format{
An object of class \code{function} of length 1.
}
\source{
\url{https://github.com/SingerLab/gac"}
}
\usage{
data(ploidyCol)
}
\description{
ploidyCol is a ComplexHeatmap scale to use when plotting ploidy phenotypes.
}
\keyword{datasets}
