% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cn_correlations.R
\name{plot_cn_correlations}
\alias{plot_cn_correlations}
\title{plot genome wide correlations}
\usage{
plot_cn_correlations(cnr, ...)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{...}{additional parameters passed to \code{Heatmap}}
}
\description{
plot genome wide correlations
}
