% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_reg_histology.R
\name{plot_effect}
\alias{plot_effect}
\title{Genome-wide effects plot}
\usage{
plot_effect(
  cnr,
  effect.column = "WD-LL.vs.WD-SCL.lr.Effect.at_Lambda.s0.01",
  type = "h",
  ylab = expression(hat(italic(Beta))),
  xlab = "HSA Genome",
  ...
)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{effect.column}{character, name of the effect column to plot}

\item{type}{character, plot type, defaults to "h". See \link[graphics]{plot}}

\item{ylab}{character, y-axis label}

\item{xlab}{character, x-axis lable}

\item{...}{additional arguments passed to plot}
}
\value{
base graphics histogram like plot containing the estimated genome-side
genotypic effects.
}
\description{
Plot estimated effects genome wide.
}
\examples{
data(cnr)

cnr <- histo_logit(cnr, trait = "binary1",
   pheno0 = 0, pheno1 = 1)

plot_effect(cnr, effect.column = "0.vs.1.lr.Effect.at_Lambda.min")

}
