% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_frequencies.R
\name{plot_frequencies}
\alias{plot_frequencies}
\title{plot genome-wide amplification and deletion frequencies}
\usage{
plot_frequencies(
  cnr,
  genes = FALSE,
  cols = c("#D40000", "#318CE7"),
  xlab = "Genome",
  ylab = "Alteration Frequency",
  ylim = c(-1, 1),
  ...
)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{genes}{logical, if TRUE use AmpFQ from gene data instead of bin, Default FALSE}

\item{cols}{colors for amplifications and deletions, default to red and blue}

\item{xlab}{x-axis label, passed to plot}

\item{ylab}{y-axis label, passed to plot}

\item{ylim}{sets coordinate space for y-axis, defaults to -1 to 1.  The
text will show absolute values for the deletion frequenies}

\item{...}{other arguments to plot}
}
\value{
Returns a bin alteration frequency plot with amplifications on the top axis,
and deletions on the bottom axis.
}
\description{
plot genome-wide amplification and deletion frequencies
}
\examples{

data(cnr)
cnr <- get_alteration_frequencies(cnr)

plot_frequencies(cnr)
plot_frequencies(cnr, xaxs = "i", bty = "n")

}
