% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_reg_histology.R
\name{plot_lr}
\alias{plot_lr}
\title{Manhattan plot of logistic regression output}
\usage{
plot_lr(
  cnr,
  pval = "LL.vs.SCL.lr.p.value",
  sig.threshold = 10^-8,
  pch = 21,
  cex = 0.64,
  ylab = expression(-Log[10] ~ P - value),
  xlab = "HSA Genome",
  ...
)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{pval}{character, name of the column in the phenotype matrix (Y)
containing p-values}

\item{sig.threshold}{numeric, significance threshold to use}

\item{pch}{plot character, default 21}

\item{cex}{character, expansion, default 0.64}

\item{ylab}{character, y-axis label}

\item{xlab}{character, x-axis label}

\item{...}{additional arguments passed to `plot`}
}
\value{
A manhattan plot with p-values for e.g. logistic regresion or other comparisons
}
\description{
Manhattan plot of logistic regression output
}
