% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCL.R
\name{plot_sK}
\alias{plot_sK}
\title{plot kCC vs sK
plot kParameter, and stable K from kStats}
\usage{
plot_sK(
  cnr,
  kCC = NULL,
  sK = NULL,
  pch = 21,
  bg = "gray",
  xaxs = "i",
  yaxs = "i",
  bty = "l",
  xpd = TRUE,
  xlab = "k Parameter (kCC)",
  ylab = "k Stable (kS)",
  segment.col = "#005EB8",
  segment.spacing = -0.8,
  segment.lwd = 1.8,
  highlight.col = "#282828",
  highlight.cex = 1.1,
  highlight.pch = 16,
  ...
)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{kCC}{kCC, when NULL, kCC is taken from `optK`}

\item{sK}{sK, when NULL kS is taken from optK}

\item{pch}{pch for all points, default is 21}

\item{bg}{background color for all points, default is "gray"}

\item{xaxs}{x axis parameter see \code{\link{par}}, default is "i"}

\item{yaxs}{y axis parameter see \code{\link{par}}, default is "i"}

\item{bty}{boxtype \code{\link{par}}, default is "l"}

\item{xpd}{allow margin text margin text \code{\link{par}}, default is TRUE}

\item{xlab}{x axis label, default is "k Parameter (kCC)"}

\item{ylab}{y axis label, default is "k Stable (kS)"}

\item{segment.col}{segment line color, default is Scotish blue "#005EB8"}

\item{segment.spacing}{segment spacing before highighted point, default is -0.8}

\item{segment.lwd}{segment line width, default 1.8}

\item{highlight.col}{color of a kCC /kS point, default is black "#282828"}

\item{highlight.cex}{character expansion of a kCC /kS point, default is 1.1}

\item{highlight.pch}{plot character of a kCC /kS point, default is 16}

\item{...}{other parameters passed to plot}
}
\description{
plot kCC vs sK
plot kParameter, and stable K from kStats
}
