% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCL.R
\name{plotCL}
\alias{plotCL}
\title{plot cluster optimization}
\usage{
plotCL(optC, type = "b", ...)
}
\arguments{
\item{optC}{output of optClust()}

\item{...}{additional arguments passed to plot and points}
}
\description{
plot cluster optimization
}
\examples{

data(cnr)

cnr <- phyloCNR(cnr)

( mopc <- optClust(cnr, opt.range = seq(0.01, 0.2, by = 0.005)) )

plotCL(optC = mopc)

}
