% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_diversity.R
\name{proportion_of_polymorphic_loci}
\alias{proportion_of_polymorphic_loci}
\title{proportion of polymorphic alleles as subclonal}
\usage{
proportion_of_polymorphic_loci(
  cnr,
  exclude.chr = c("X", "Y"),
  monomorphic.threshold = 0.95,
  noise.threshold = 0.1
)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{exclude.chr}{chromosomes to exclude, default X, Y}

\item{monomorphic.threshold}{alteration frequency at which a
locus is to be considered monomorphic (universally altered).
default 0.95}

\item{noise.threshold}{lower-bound alteration frequency at
which alterations are considered technical noise, rather
than a true alteration, default 0.1}
}
\value{
Single-value with the proportion of polymorphic loci for a given
sample population
}
\description{
Estimate genome-wide proportion of polymorphic loci in a population
}
\examples{
data(cnr)

cnr <- get_alteration_frequenciesCNR(cnr)

proportion_of_polymorphic_loci(cnr)

}
