% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClusters.R
\name{rank_clones}
\alias{rank_clones}
\title{rank clones}
\usage{
rank_clones(cnr, cluster_column, rank_by, fga.method, prefix = NULL)
}
\arguments{
\item{cnr}{a cnr object}

\item{cluster_column}{which cluster column to use; must be one of
"BrayC", "ConsensusC", or "final_cluster"}

\item{rank_by}{option to rank, 'fga' to rank by genomic complexity based
on the fraction of genome altered. Or 'frequency', ranked from most to
least least prevalent.}

\item{fga.method}{option to estimate cluster FGA, either 'genomic' or 'proxy'.
In method 'genomic' the altered bins are weighed by their genomic length in
base pairs (bp), and the fraction is estimated based on the total genome
length.  Method 'proxy' does not weigh by genomic lenght, instead it
approximates FGA by estimating the proportion of altered bins.  Alterations
are deviations from a diploid state e.g. != 2.}

\item{prefix}{prefix charcter to append to Consensus Clusters}
}
\value{
rank_clones returns a ranked_cluster vector to `Y`.  The ranking of the clone is
user specified by frequency or genomic complexity.  When ranking clones by frequency,
the specified clusters are ranked in descending order, where the most frequent clone
is assigned clone `1`, and all less frequent clones are assigned a numerical label
in sequential order.  Ties are broke at random.  When ranking by genomic complexity,
then the fraction of the genome altered (FGA) is computed based on a diploid genome (2n).
And ranked in ascending order, where the least altered cell is clone `1` and more altered
clones are assigned a numerical label in sequential order.  Ties are broken at random.

Clusters with 3 or more cells are sumarized using the same parameters as in \link{get_cluster_profiles}.
However, for clusters with 2 cells, are sumarized by taking the `max` at each bin, and those with one cell, then
the cell was used.
}
\description{
STILL IN DEVELOPMENT: UNTESTED
}
