% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_seg_to_bins.R
\name{rle_ddrc}
\alias{rle_ddrc}
\title{run length encoding to create segment files from DDRC.df}
\usage{
rle_ddrc(cnr, chrom = "bin.chrom", pos = "bin.start")
}
\arguments{
\item{cnr}{a cnr bundle}

\item{chrom}{name of column in chromInfo to use as chromosomes}

\item{pos}{name of column in chromInfo to use as chromosomal position}
}
\value{
A run length encoding table from the clone profiles contained
in DDRC.df
}
\description{
Still in development:  currently performs RLE by chromosome and
returns chromsome specific encoding.  Locantions are not currently 
offsest by chromsome start locations
}
\examples{

data(cnr)

noisy.cells <- cnr$qc$cellID[cnr$qc$qc.status == "FAIL"]

## reduced pipeline to genrate DDRC clone profiles
cnr <- excludeCells(cnr, excl = noisy.cells)
cnr <- phyloCNR(cnr, root.cell = "cell0")
cnr <- setBrayClusters(cnr)
cnr <- run_consensus_clustering(cnr, iters = 20, maxK = 40)
cnr <- doKSpectral(cnr)
cnr <- setKcc(cnr)
cnr <- cluster_heterogeneity(cnr, by = "category1",
          cluster_column = "ConsensusC")
cnr <- get_cluster_profiles(cnr)

ddrc_rle <- rle_ddrc(cnr)

}
