% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_consensus_clustering.R
\name{run_consensus_clustering}
\alias{run_consensus_clustering}
\title{run consensus clustering}
\usage{
run_consensus_clustering(
  cnr,
  maxK = 40,
  iters = 200,
  title = "cnr_ccp",
  innerLinkage = "ward.D2",
  finalLinkage = "ward.D2",
  seed = 2020.0314,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{cnr}{a cnr object}

\item{maxK}{maximum K, end of K sequence}

\item{iters}{number of times the clustering runs}

\item{title}{title for the consensus cluster map}

\item{innerLinkage}{inner linkage, see ConsensusClusterPlus}

\item{finalLinkage}{final linkage, see ConsensusClusterPlus}

\item{seed}{seed of analysis}

\item{verbose}{print out progress}

\item{...}{additional parameters for ConsensusClusterPlus e.g. plot = "png"}
}
\value{
A cnr object that contains the output of ConsensusClusterPlus for the
 specified Ks.

\itemize{
  \item ccp ConsensusClusterPlus object
}
}
\description{
Function runs consensus clustering for a pre-determined sequence of K.
}
\examples{
data(cnr)

cnr <- phyloCNR(cnr)

cnr <- run_consensus_clustering(cnr, maxK = 6)

}
