% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scClustCNR.R
\name{scClustCNR}
\alias{scClustCNR}
\title{implementation of SCclust pipeline}
\source{
\url{https://github.com/KrasnitzLab/SCclust}
}
\usage{
scClustCNR(
  cnr,
  cytobands,
  hc.method = "average",
  tree.method = "average",
  run.fisher = TRUE,
  ...
)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{cytobands}{UCSC cytobands file}

\item{hc.method}{clustering method for hclust_tree}

\item{tree.method}{clustering method for tree_py}

\item{run.fisher}{logical, weather you want to run sim_fisherCNR. defaults TRUE.
If FALSE, it assumes you have ran sim_fisherCNR before, and the cnr already
contains cnr[["pins"]] and cnr[["fisher"]]}

\item{...}{additional arguments passed}
}
\description{
implementation of SCclust pipeline
}
\keyword{internal}
