% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_seg_to_bins.R
\name{seg2bins}
\alias{seg2bins}
\title{convert .seg files to match a chromInfo matrix}
\usage{
seg2bins(
  cnr,
  seg,
  sample.id = "ID",
  bin.id = NULL,
  coordinates.a = c("bin.chrom", "bin.start", "bin.end"),
  coordinates.b = c("chrom", "loc.start", "loc.end"),
  seg.mean.b = "seg.mean",
  multipcf = TRUE
)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{seg}{a table in .seg format}

\item{sample.id}{column name of sample ID in the .seg dat, default is "ID"}

\item{bin.id}{column name containing a bin.id, defalt NULL,
bin.id will be taken as the rownames of the data}

\item{coordinates.a}{names of the columns containing chromosome, start,
and end coordinates in the cnr}

\item{coordinates.b}{names of the columns containing chromosome, start,
and end coordinates in the seg data}

\item{seg.mean.b}{name of column containg segment means default "seg.mean"}

\item{multipcf}{weather to create a multi sample matrix of the .seg data
or keep as run length encoding, akin to copynumber::multipcf,
and copynumber::pcf, respectively.  This function requires
tidyr::pivot_wider}
}
\description{
convert .seg files to match a chromInfo matrix
}
\examples{
\dontrun{
data(cnr)
seg.data <- read.delim("path/to/data.seg")

binned.seg <- seg2bins(cnr, seg.data,
                       sample.id = "sampleID",
                       bin.id = "bin.id")
}

}
