% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segCol.R
\docType{data}
\name{segCol}
\alias{segCol}
\title{segCol -- ComplexHeatmap color scale}
\format{
An object of class \code{function} of length 1.
}
\source{
\url{https://github.com/SingerLab/gac/"}
}
\usage{
data(chromInfo)
}
\description{
segCol is a ComplexHeatmap color scale to use when plotting X data
}
\keyword{datasets}
