% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClusters.R
\name{setBrayClusters}
\alias{setBrayClusters}
\title{setting clusters at optimal height}
\usage{
setBrayClusters(cnr, tree.height, prefix = "C")
}
\arguments{
\item{cnr}{a cnr bundle}

\item{tree.height}{height of the tree (from optClust)}

\item{prefix}{a prefix to append before the cluster number}
}
\value{
Returns a cnr object with cluster membership based on Bray-Curtis
 heirarchical clustering.  Tree height must be specified.
}
\description{
setting clusters at optimal height
}
