% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClusters.R
\name{setClusters}
\alias{setClusters}
\title{setting clusters at optimal height}
\usage{
setClusters(cnr, tree.height, prefix = "C")
}
\arguments{
\item{cnr}{a cnr bundle}

\item{tree.height}{height of the tree (from optClust)}

\item{prefix}{a prefix to append before the cluster number}
}
\description{
setting clusters at optimal height
}
