% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setKcc.R
\name{setKcc}
\alias{setKcc}
\title{pull cluster membership for K clusters}
\usage{
setKcc(cnr, K = 5)
}
\arguments{
\item{cnr}{a cnr object}

\item{K}{number of K clusters to use}
}
\value{
returns cluster membership based on consensus clustering for a
 specified K
}
\description{
pull cluster membership for K clusters
}
