% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_cnr.R
\name{split_cnr}
\alias{split_cnr}
\title{split a cnr by one variable}
\usage{
split_cnr(cnr, split.by)
}
\arguments{
\item{cnr}{a cnr object}

\item{split.by}{name of the column to split. Must be a column in
the phenotype or qc tables}
}
\value{
Function subsets based on `split.by` and returns a list of named cnr objects.
 Element names are taken from the unique values of split.by
}
\description{
split a cnr by one variable
}
\examples{

data(cnr)

cnrL <- split_cnr(cnr, split.by = "category1")

lapply(cnrL, summary_cnr)

}
