% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_diversity.R
\name{subset_ci}
\alias{subset_ci}
\title{subset for chromosomes of interest}
\usage{
subset_ci(cnr, exclude.chr = NULL, chrom.col = "bin.chrom")
}
\arguments{
\item{cnr}{a cnr bundle}

\item{exclude.chr}{vector, chromosomes to exclude}

\item{chrom.col}{name of column containing chromosomes}
}
\value{
conditional subset of chromInfo w/o excluded chromosomes
}
\description{
subset for chromosomes of interest
}
\keyword{internal}
