% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetCNR.R
\name{subset_on_bins}
\alias{subset_on_bins}
\title{subset based on bins}
\usage{
subset_on_bins(cnr, bins)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{bins}{a list of bins to subset}
}
\description{
subset based on bins
}
\keyword{internal}
