% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetCNR.R
\name{subset_on_chrom}
\alias{subset_on_chrom}
\title{subset based on complete chromosomes}
\usage{
subset_on_chrom(cnr, chrom)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{chrom}{a list of chromosomes of interst}
}
\description{
subset based on complete chromosomes
}
\keyword{internal}
