% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetCNR.R
\name{subset_on_coordinates}
\alias{subset_on_coordinates}
\title{subset based on complete chromosomes}
\usage{
subset_on_coordinates(cnr, chrom, start, end, padding = 1)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{chrom}{a chromosome, only one chromosome}

\item{start}{start coordinate}

\item{end}{end coordinate}

\item{padding}{number of bins to add before or after the start and end coordinates}
}
\description{
subset based on complete chromosomes
}
\keyword{internal}
