% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetCNR.R
\name{subset_on_genes}
\alias{subset_on_genes}
\title{subset based on genes}
\usage{
subset_on_genes(cnr, genes, all = TRUE)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{genes}{a list of genes to subset}

\item{all}{return all genes within on the bins of the genes of interst}
}
\description{
subset based on genes
}
