% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetCNR.R
\name{subsetCNR}
\alias{subsetCNR}
\title{Subset a set of bins or genes from a CNR object}
\usage{
subsetCNR(
  cnr,
  bins = NULL,
  genes = NULL,
  chrom = NULL,
  start = NULL,
  end = NULL,
  ...
)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{bins}{a vector of bins to keep}

\item{genes}{a vector of genes to keep}

\item{chrom}{a vector of chromosomes, or single chromosome if genomic coordinates are used}

\item{start}{a start position}

\item{end}{an end position}

\item{...}{additional parameters e.g. padding to add bins when coordinates are used}
}
\value{
Returns a cnr object with only the desired bins.  This function also subsets the gene index and chromInfo as well
}
\description{
This function pulls out a section of the genome for further analysis.
}
\examples{
data(cnr)
data(segCol)

## subset based on bins
bins1.100 <- subsetCNR(cnr, bins = 1:100)

## subset based on genes
genes.3 <- subsetCNR(cnr, genes = c("CDK4", "MDM2", "HMGA2", "TP53"))

## subset based on chromosomes
chrom6.12 <- subsetCNR(cnr, chrom = c(6, 12))

## subset based on genomic coordinates
chrom12q13.15 <- subsetCNR(cnr, chrom = 12, start = 46000000, end = 71000000)

}
