% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryCNR.R
\name{summary_cnr}
\alias{summary_cnr}
\title{summary of cnr bundle}
\usage{
summary_cnr(cnr, detailed = FALSE, display = FALSE)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{detailed}{logical, to provide additional detail on the bundle. Default (FALSE)}

\item{display}{logical, weather to display data and other results. Default (FALSE)}
}
\value{
A summary of the contents in the cnr
}
\description{
summary of cnr bundle
}
\examples{

data(cnr)

summary_cnr(cnr)

}
