% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kSpectral.R
\name{symmetricSinkhornKnopp}
\alias{symmetricSinkhornKnopp}
\title{symmetricSinkhornKnopp}
\usage{
symmetricSinkhornKnopp(A, tol = NA, maxiter = Inf, debug = FALSE)
}
\arguments{
\item{A}{symmetric matrix}

\item{tol}{tol default is NA, which computes `tol = eps(N)`}

\item{maxiter}{maximum number of iterations, default is Inf}

\item{debug}{logical, turn on/off debugging i.e. TRUE/FALSE; default is FALSE}
}
\description{
symmetricSinkhornKnopp
}
\references{
Philip A. Knight (2008) The Sinkhorn–Knopp Algorithm: Convergence and Applications. SIAM Journal on Matrix Analysis and Applications 30(1), 261-275. doi: 10.1137/060659624
}
\author{
ND Socci <soccin .at. mskcc dot org>
}
\keyword{internal}
