% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cluster_profiles.R
\name{use_clusters}
\alias{use_clusters}
\title{select clusters to use based on a minimum number of cells}
\usage{
use_clusters(cnr, minimum_cells = 3)
}
\arguments{
\item{cnr}{a cnr bundle containing the output of setCluster}

\item{minimum_cells}{minimum number of cells in a cluster, must be greater than 3
to estimate a median}
}
\description{
select clusters to use based on a minimum number of cells
}
\keyword{internal}
