% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdjBrayClust.R
\name{vdjBrayClust}
\alias{vdjBrayClust}
\title{VDJ cells clustering based on Bray-Curtis Dissimilarity}
\usage{
vdjBrayClust(cnr, vdj.genes = NULL)
}
\arguments{
\item{cnr}{a cnr bundle, previously genotyped for VDJ}

\item{vdj.genes}{list of VDJ genes for clustering. All must be in the
gene.index `hgnc.symbol` column}
}
\description{
VDJ cells clustering based on Bray-Curtis Dissimilarity
}
