% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdjBrayClust.R
\name{vdjPlotClust}
\alias{vdjPlotClust}
\title{plot VDJ clustering}
\usage{
vdjPlotClust(cnr, type = "fan", ...)
}
\arguments{
\item{cnr}{a cnr bundle with clustered vdj cells}

\item{type}{type of tree, from \code{\link[ape]{plot.phylo}}, default is "fan",}

\item{...}{additional parameters to \code{\link[ape]{plot.phylo}}}
}
\description{
plot VDJ clustering
}
