% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addInfo.R
\name{addGeneInfo}
\alias{addGeneInfo}
\title{Adds Gene Information to the gene.index}
\usage{
addGeneInfo(cnr, df, sort = FALSE, ...)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{df}{a data.frame with the data to incorporate.  Particularly
useful for gene annotations, filters e.g. OncoKB, or results from
analyses e.g. p-values, genetic effects, etc to the genes}

\item{sort}{wether to sort the ouput object, default is FALSE}

\item{...}{additional parameters passed to merge}
}
\value{
Returns a CNR object with added columns from df to the gene.index.
  e.g. p-values for genome wide scans
}
\description{
The function adds information to the gene.index.#'
}
\examples{
#' data(cnr)

fakePval <- data.frame(pval = runif(5000))

cnr <- addGeneInfo(cnr, df = fakePval)

head(cnr$gene.index)

}
