% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPheno.R
\name{addPheno}
\alias{addPheno}
\title{Adds additional phenotypes for a new trait}
\usage{
addPheno(cnr, df, ...)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{df}{a data frame with new Y traits}

\item{...}{additional arguments by merge}
}
\description{
add phenotype columns to the data
}
\examples{

data(cnr)

rand3 <- cbind(cnr$Y, rnorm(nrow(cnr$Y), mean = 2, sd = 1))

cnr <- addPheno(cnr, df = rand3)

}
