% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary.cnr.R
\name{binary.X}
\alias{binary.X}
\title{build binary matrix from integer copy number data}
\usage{
binary.X(X, bulk, base.ploidy = 2)
}
\arguments{
\item{X}{a copy number matrix to convert to an incidence matrix}

\item{bulk}{logical, if TRUE data is bulk DNA is cbioportal calls, default is FALSE for integer copy number}

\item{base.ploidy}{expected ploidy, e.g. 2N = 2, 4N = 4}
}
\value{
Returns an incidence matrix for X, when X != base.ploidy
}
\description{
This function builds a binary alteration incidence matrix from copy number data.
Anything not equal to base.ploidy is 1.
}
\examples{

data(cnr)

Z <- binary.X(cnr$X, bulk = cnr$bulk)

Z <- binary.X(cnr$genes[, c("CDK4", "MDM2")], bulk = cnr$bulk)

}
