% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_chromAnno.R
\name{chr_breaks}
\alias{chr_breaks}
\title{estimate chromosome end locations along a continuous genome}
\usage{
chr_breaks(cnr, bin = TRUE)
}
\arguments{
\item{cnr}{a cnr}

\item{bin}{weather to use bin or gene data, default is true}
}
\value{
A named vector of chromosome breaks locations in the data.
Useful when adding lines to seaparete chromosomes,  or
a background when highlighting a chromosome
}
\description{
estimate chromosome end locations along a continuous genome
}
