% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromInfo.R
\docType{data}
\name{chromInfo}
\alias{chromInfo}
\title{Bin chromosome-basepair coordinates}
\format{
A data frame with chromosome and position information
\describe{
  \item{chr}{chromosome}
  \item{pos}{base pair coordinate}

  ...
}
}
\source{
\url{https://github.com/SingerLab/toSignac}
}
\usage{
chromInfo
}
\description{
A dataset containing the end chromsome coordiantes for all the bins of the single-cell copy number matrix `cnq`
}
\keyword{datasets}
