% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_heterogeneity.R
\name{cluster_heterogeneity}
\alias{cluster_heterogeneity}
\title{cluster heterogeneity across the sample}
\usage{
cluster_heterogeneity(cnr, by = NULL, cluster_column = NULL)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{by}{a categorical variable used to stratify the cell population.  If
NULL i.e. no stratification, the representation will be done overall}

\item{cluster_column}{column to use as cluster}
}
\value{
A `data.frame` with a simple cluster heterogeneity summary.  Table provides total
 counts of each clone (optionally be stratified by a group), clone representation,
 overal frequency per clone (also computed for stratified data), and Shannon's
 Diversity index for stratified data.
}
\description{
cluster heterogeneity across the sample
}
\examples{

data(cnr)

cnr <- phyloCNR(cnr)

cnr <- setBrayClusters(cnr, tree.height = 0.065)

cnr <- cluster_heterogeneity(cnr, cluster_column = "BrayC")

}
