% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_heterogeneity.R
\name{cluster_heterogeneity}
\alias{cluster_heterogeneity}
\title{cluster heterogeneity across the sample}
\usage{
cluster_heterogeneity(cnr, by = NULL)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{by}{a categorical variable used to stratify the cell population.  If
NULL i.e. no stratification, the representation will be done overall}
}
\description{
cluster heterogeneity across the sample
}
\examples{

data(cnr)

cnr <- phyloCNR(cnr)

cnr <- setClusters(cnr, tree.height = 0.065)

cnr <- cluster_heterogeneity(cnr)


}
