% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_chromAnno.R
\name{create_chromosome_annotation}
\alias{create_chromosome_annotation}
\title{create chromosome annotations for custom heatmaps}
\usage{
create_chromosome_annotation(cnr, side = "left", ...)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{side}{side of the annotation. Options are "left" default,
"right", "top", and "bottom"}

\item{...}{additional prameters passed to HeatmapAnnotation}
}
\value{
A HeatmapAnnotation containing a chromosome map indicate chromsome
boundaries on heatmaps
}
\description{
create chromosome annotations for custom heatmaps
}
\examples{
data(cnr)

chrAnnoLeft <- create_chromosome_annotation(cnr)
chrAnnoTop <- create_chromosome_annotation(cnr, side = "top")

}
