% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_chromAnno.R
\name{create_chromosome_annotation_bottom}
\alias{create_chromosome_annotation_bottom}
\title{create chromosome annotations for custom heatmaps}
\usage{
create_chromosome_annotation_bottom(
  cnr,
  labels_gp = grid::gpar(fontsize = 10),
  labels_rot = 0,
  ...
)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{labels_gp}{graphic parameters from \link[grid]{gpar}, default fontsize = 10}

\item{labels_rot}{label rotation, default 90}

\item{...}{additional prameters passed to HeatmapAnnotation}
}
\description{
create chromosome annotations for custom heatmaps
}
