% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_frequencies.R
\name{get_bin_frequencies}
\alias{get_bin_frequencies}
\title{estimate bin alteration frequencies}
\usage{
get_bin_frequencies(cnr, ...)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{...}{additional parameters passed to get_amp_counts, get_del_counts, or
get_alt_counts}
}
\description{
estimate bin alteration frequencies
}
\keyword{internal}
