% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_frequencies.R
\name{get_gene_frequencies}
\alias{get_gene_frequencies}
\title{estimate bin alteration frequencies}
\usage{
get_gene_frequencies(cnr)
}
\arguments{
\item{cnr}{a cnr bundle}
}
\description{
estimate bin alteration frequencies
}
