% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisticGR.R
\name{gisticGR}
\alias{gisticGR}
\title{Gistic Genomic Ranges, get genes in gene.index that overlap with significant genomic ranges in GISTIC2 output}
\usage{
gisticGR(cnr, grTR, oncokb, depmeans, cutoff = -0.5)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{grTR}{gistic regions table, output from gisticRegions}

\item{oncokb}{OncoKB annotation (downloaded from OncoKB)}

\item{depmeans}{dependency score means for cells of interest
e.g. DepMap achilles scores}

\item{cutoff}{threshold to consider a gene as essential or non-essential}
}
\description{
**Still in development**
** requires GenommicRanges
}
