% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisticGR.R
\name{gisticGR}
\alias{gisticGR}
\title{Gistic Genomic Ranges, get genes in gene.index that overlap with significant
genomic ranges in GISTIC2 output}
\usage{
gisticGR(cnr, grTR, oncokb = TRUE, depmeans = TRUE)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{grTR}{gistic regions table, output from gisticRegions}

\item{oncokb}{OncoKB cancer gene annotations (TRUE/FALSE)
(see ?oncoKB_annotation)}

\item{depmeans}{dependency score means for cells of interest
e.g. DepMap achilles scores (see ?depmap_annotation), (TRUE/FALSE)}
}
\description{
**Still in development**
** requires GenommicRanges
}
\details{
note if oncokb or depmeans are FALSE, both annotations will be ignored
}
\keyword{internal}
