% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupCN.R
\name{lookupCN}
\alias{lookupCN}
\title{Lookup copy number data from bins}
\usage{
lookupCN(cnr, coord)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{coord}{a `list` with elements named chr start end}
}
\value{
Returns the X data for a set of chromosome coordinates
}
\description{
This function pulls the X copy number data based on coordinate information from the cnr$chromInfo
}
\examples{

data(cnr)
coords <- list(chr = 2, start = 550000, end = 600000)

lookupCN(cnr, coord = coords)

}
