% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupCN.R
\name{lookupCN}
\alias{lookupCN}
\title{Lookup copy number data from bins}
\usage{
lookupCN(cnr, chr = 12, start = 69200804, end = 69246466)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{chr}{a chromosome name, must match 'cnr$chromInfo$bin.chrom'}

\item{start}{start coodinate}

\item{end}{end coordinate}
}
\value{
Returns the X data for a set of chromosome coordinates
}
\description{
This function pulls the X copy number data based on coordinate information from the cnr$chromInfo
}
\examples{

data(cnr)
coord <- data.frame(chr = 2, start = 550000, end = 600000)

lookupCN(cnr, chr = coord$chr, start = coord$start, end = coord$end)

}
