% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_chromAnno.R
\name{mid_chr}
\alias{mid_chr}
\title{estiamte chromosome midpoint locations along a continuous genome}
\usage{
mid_chr(cnr, bin = TRUE)
}
\arguments{
\item{cnr}{a cnr}

\item{bin}{weather to use bin or gene data, default is true}
}
\value{
A named vector of chromosome midpoints. Useful for adding tick
marks in figures.  Midpoint is not the centromere location.
}
\description{
estiamte chromosome midpoint locations along a continuous genome
}
