% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryCNR.R
\name{nbins}
\alias{nbins}
\title{number of bins}
\usage{
nbins(cnr, detailed = FALSE, display = FALSE)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{detailed}{detailed summary}

\item{display}{logical, display chromInfo}
}
\description{
number of bins
}
