% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_cnr.R
\name{ncells}
\alias{ncells}
\title{number of cells in cnr}
\usage{
ncells(cnr, detailed = FALSE, detailed.column = NULL, display = FALSE)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{detailed}{logical, detailed summary e.g print total cells, n vdj cells, & cells by cell type}

\item{detailed.column}{character, name of column to partition cells, default is NULL}

\item{display}{display cell ID's}
}
\description{
number of cells in cnr
}
\examples{
data(cnr)

ncells(cnr)

ncells(cnr, detailed = TRUE)

ncells(cnr, detailed = TRUE,
       detailed.column = "category1")

}
