% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryCNR.R
\name{ngenes}
\alias{ngenes}
\title{number of genes in gene.index}
\usage{
ngenes(cnr, detailed = FALSE, display = FALSE)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{detailed}{detailed summary}

\item{display}{logical, display gene.index}
}
\description{
number of genes in gene.index
}
