% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optClust.R
\name{optClust}
\alias{optClust}
\title{optimizing clustering for single-cell copy number}
\usage{
optClust(cnr, opt.range = seq(0.005, 0.6, by = 0.005))
}
\arguments{
\item{cnr}{a cnr bundle}

\item{opt.range}{range of tree heights that need to be optimized}
}
\value{
Returns a matrix where rows are the heights in the opt.range, and three columns.

One-cell specifies the number of clusters with only one cell, 
Multi-cell specifyies the number of clusters with multiple cells (>=2), and
%CMC which is the percentage of cells in multi-cell clusters

A good threshold is one that minimizes the one-cell cluster, and maximizes
the %CMC
}
\description{
The optClust is a simple attempt to set a threshold on where to cut the an
hclust tree.  The aim is to maximize the number of cluster with multiple-cells
while minimizing the number of one-cell clusters.
}
\examples{

data(cnr)

cnr <- phyloCNR(cnr)

optClust(cnr, opt.range = seq(0, 0.3, by = 0.05))

#for micro-optimization

optClust(cnr, opt.range = seq(0, 0.2, by = 0.001))

}
