% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloCNR.R
\name{phyloCNR}
\alias{phyloCNR}
\title{Calculating cell-to-cell distances, heirarchical clustering, and generating a tree class `phylo`}
\usage{
phyloCNR(
  cnr,
  root.cell = NULL,
  dist.method = "bray",
  hclust.method = "ward.D2",
  ...
)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{root.cell}{a cellID to root the three.}

\item{dist.method}{method for calculating cell-to-cell distance
(see vegan::vegdist)}

\item{hclust.method}{method for heirarchical clustering (see hclust)}

\item{...}{other parameters passed to \link[vegan]{vegandist}}
}
\value{
Creates a cell-to-cell distance matrix, runs heirarchical clustering,
 and converts the `hclust` object to an `ape` class `phylo` object.

\itemize{
  \item cdb cell to cell Bray-Curtis dissimiarly 
  \item hcdb heirarchical clustering of distance matrix
  \item phylo ape class `phylo` object
}
}
\description{
Calculating cell-to-cell distances, heirarchical clustering, and generating a tree class `phylo`
}
