% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloCNR.R
\name{phyloCNR}
\alias{phyloCNR}
\title{calculating distances, heirarchical clustering, and `phylo` object}
\usage{
phyloCNR(cnr, dist.method = "bray", hclust.method = "ward.D2", ...)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{dist.method}{method for calculating cell-to-cell distance
(see vegan::vegdist)}

\item{hclust.method}{method for heirarchical clustering (see hclust)}

\item{...}{other parameters}
}
\description{
calculating distances, heirarchical clustering, and `phylo` object
}
