% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plog.R
\name{plog}
\alias{plog}
\title{internal}
\usage{
plog(x)
}
\arguments{
\item{x}{vector or matrix of copy numbers}
}
\value{
Returns plog transformed data
}
\description{
calculates a plog for parctan()
}
\examples{
data(cnr)

plog(cnr$X[,1])

}
