% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_gene_details.R
\name{pull_gene_details}
\alias{pull_gene_details}
\title{Pull gene details for a genomic region}
\usage{
pull_gene_details(cnr, coord = "12:69200804:69246466")
}
\arguments{
\item{cnr}{a cnr bundle}

\item{coord}{genomic region in ensembl format}
}
\value{
Returns the subset of the `gene.index` table for the genomic region.
}
\description{
This function subsets the gene index for a genomic region of interest.
}
\examples{

library(dplyr)
data(cnr)
coord <- "12:69200804:69246466"

pull_gene_details(cnr, coord = coord)

coords <- c("1:170120554:172941951",
          "12:69200804:69246466")

do.call(rbind, lapply(coords, function(rr)
                     pull_gene_details(cnr, coord = rr)))

}
