% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClusters.R
\name{setBrayClusters}
\alias{setBrayClusters}
\title{Set clusters using one-cell and multi-cell height intersection}
\usage{
setBrayClusters(cnr, tree.height = NULL, prefix = "C", ...)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{tree.height}{height of the tree \link{optClust}}

\item{prefix}{a prefix to append before the cluster number}

\item{...}{additional parameters passed to \link{optClust}}
}
\value{
Returns a cnr object with cluster membership based on Bray-Curtis
 Ward.D2 heirarchical clustering. If hieght is NULL, then the
 the tree height is calculated using \link{minimum.intersect}. Where
 optimal height is the value at the intersect between the number of
 one-cell and multi-cell clusters.
}
\description{
Set clusters using one-cell and multi-cell height intersection
}
