% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scClustCNR.R
\name{sim_fisherCNR}
\alias{sim_fisherCNR}
\title{implementation of SCclust sim_fisher_wrapper}
\source{
\url{https://github.com/KrasnitzLab/SCclust}
}
\usage{
sim_fisherCNR(cnr, cytobands, centromere = c("p11", "q11"), nsim = 200, ...)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{cytobands}{UCSC cytobands file}

\item{centromere}{cytoband boundaries of the centromere (will be excluded)}

\item{nsim}{number of Fisher wrapper simulations}

\item{...}{additional parameters pased to SCclust::sim_fisher_wrapper}
}
\value{
SCclust pipeline and objects
}
\description{
scClust uses Fisher distance
}
\examples{

\dontrun{
## the simulated data does not run well
data(cnr, hg19_cytoBand)
cnr <- sim_fisherCNR(cnr, cytobands = hg19_cytoBand)
}

}
\keyword{internal}
