% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetCNR.R
\name{subsetCNR}
\alias{subsetCNR}
\title{subset a set of bins or genes from a CNR object}
\usage{
subsetCNR(cnr, based.on = "X", keep, ...)
}
\arguments{
\item{cnr}{a cnr bundle}

\item{based.on}{"X", "genes", "position"}

\item{keep}{string vector with which bins or genes to keep}

\item{...}{additional parameters}
}
\value{
Returns a cnr object with only the desired bins.  This function also subsets the gene index and chromInfo as well
}
\description{
This function pulls out a section of the genome for further analysis.
}
\details{
** still in development **


It currently only works based on bin data, but working to get gene based subset and chromsome position subset.
}
\examples{
data(cnr)
data(segCol)

chr19 <- which(cnr$chromInfo$chr == "19")
cnr_chr19 <- subsetCNR(cnr, based.on = "X", keep = chr19)

sapply(cnr_chr19, dim)

HeatmapCNR(cnr_chr19, col = segCol)

}
