% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toSignac.R
\docType{package}
\name{toSignac}
\alias{toSignac}
\title{toSignac: Sinple integrated analysis of cells/copynumber}
\description{
Pronouced: t o Signac , as in the open-source Signac
}
\details{
toSignac is an open-source framework and toolkit for managing genome wide copy number data.  It provides a set of standard matrices that are symultaneously manipulated by various functions to maintain syncronized bin, gene, annotation, and qc information from a single-cell DNA sequencing experiment as same-cell G+T.  Efforts will be made to add support for log-ratios for the analysis of bulk DNA data.


toSignac is actively being developed and is currently in ALPHA pre-release.  My hope is that others will find this helpful for their copy number needs, and share new methods.

The HeatmapCNR function in toSignac implements Bray-Curtis disimilarity on ComplexHeatmap as the default method to estimate cell-cell distances and generate the heatmaps
}
\author{
Rodrigo Gularte Merida \email{gularter@mskcc.org}
}
