% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdjHeatmap.R
\name{vdjHeatmap}
\alias{vdjHeatmap}
\title{Heatmap of cells with VDJ recombination}
\usage{
vdjHeatmap(
  cnr,
  vdj.genes = NULL,
  vdjGeneAnno = NULL,
  vdjCellAnno = NULL,
  gene.type.column = "gene_biotype",
  ...
)
}
\arguments{
\item{cnr}{a cnir bundle}

\item{vdj.genes}{list of VDJ genes.  By default is set to NULL, vdj.genes is
the union of T-cell Receptor and B-cell Receptor genes (i.e. TR, IG genes,
used in genotype_vdj)}

\item{vdjGeneAnno}{HeatmapAnnotation object to annotate gene attributes}

\item{vdjCellAnno}{rowAnnotation object to annotate cells}

\item{gene.type.column}{name of column with gene type, default "gene_biotype}

\item{...}{additional arguments passed to ComplexHeatmap::Heatmap}
}
\value{
Returns a heatmap of VDJ cells with rows as cells, and TR/IG genes as columns.
 Heatmap is split by chromosome name, and cell types are annotated on the left
}
\description{
Heatmap of cells with VDJ recombination
}
\examples{
data(cnr)

cnr <- genotype_vdj(cnr)

vdjHeatmap(cnr)

}
